Crate gooey

A graphical user interface library. This crate exposes all of the built-in functionality of Gooey, as well as types for building apps.

Feature Flags

This crate has several feature flags to control what features are enabled. The default feature flags are ["frontend-kludgine", "frontend-browser", "fluent"].

frontend-browser: Enables the frontends::browser module, which is gooey-browser re-exported.

frontend-kludgine: Enables the frontends::rasterizer module and the frontends::renderers::kludgine module. These are re-exports of gooey-rasterizer and gooey-kludgine respectively.

async: Enables the App::spawn() function.

fluent: Enables using fluent for localization.

Top-level exports

The ActiveFrontend, main(), and main_with() function will change types based on the feature flags enabled as well as the target platform. Here is a reference to understand what symbols to expect:

target_arch = "wasm32"	frontend-kludgine	frontend-browser	ActiveFrontend/main()

false	false	false

false	false	true

false	true	false	Rasterizer<Kludgine>/kludgine_main

false	true	true	Rasterizer<Kludgine>/kludgine_main

true	false	true	WebSys/browser_main

true	true	true	WebSys/browser_main

Modules

core
Core traits and types used to create Graphical User Interfaces (GUIs - gooeys).

fluent
Localization support for Gooey using Project Fluent.

frontends
Available Frontends.

renderer
Cross-platform rendering types.

style
Styles for applications.

text
Text rendering and wrapping.

widgets
Built-in widgets for user interfaces.

Structs

App
A cross-platform application.

Headless
A headless application.

Recorder
An easy-to-use,offscreen animation recorder.

Enums

HeadlessError
Errors that can occur while using Headless.

Functions

app
Returns an initialized frontend using the root widget returned from initializer.

browser_app
Returns an initialized frontend using the root widget returned from initializer.

browser_main
Runs a browser-based App with the root widget from initializer. All widgets from gooey::widget will be usable. If you wish to use other widgets, use browser_main_with and provide the transmogrifiers for the widgets you wish to use.

browser_main_with
Runs a browser-based App with transmogrifiers and the root widget from initializer. Unless overriden by transmogrifier, all widgets from gooey::widget will use the built-in transmogrifiers.

browser_run
Runs an initialized frontend.

kludgine_app
Returns an initialized frontend using the root widget returned from initializer.

kludgine_main
Runs a Kludgine-based App with the root widget from initializer. All widgets from gooey::widget will be usable. If you wish to use other widgets, use browser_main_with and provide the transmogrifiers for the widgets you wish to use.

kludgine_main_with
Runs a Kludgine-based App with transmogrifiers and the root widget from initializer. Unless overriden by transmogrifier, all widgets from gooey::widget will use the built-in transmogrifiers.

kludgine_run
Runs an initialized frontend.

main
Runs a Kludgine-based App with the root widget from initializer. All widgets from gooey::widget will be usable. If you wish to use other widgets, use browser_main_with and provide the transmogrifiers for the widgets you wish to use.

main_with
Runs a Kludgine-based App with transmogrifiers and the root widget from initializer. Unless overriden by transmogrifier, all widgets from gooey::widget will use the built-in transmogrifiers.

run
Runs an initialized frontend.

Type Definitions

ActiveFrontend
The active frontend.