Module gooey::animation

Types for creating animations.

Animations in Gooey are performed by transitioning a Dynamic’s contained value over time. This starts with Dynamic::transition_to(), which returns a DynamicTransition.

DynamicTransition implements AnimationTarget, a trait that describes types that can be updated using linear interpolation. AnimationTarget is also implemented for tuples of AnimationTarget implementors, allowing multiple transitions to be an AnimationTarget.

Next, the AnimationTarget is turned into an animation by invoking AnimationTarget::over() with the Duration the transition should occur over. The animation can further be customized using Animation::with_easing() to utilize any Easing implementor.

use std::time::Duration;

use gooey::animation::easings::EaseInOutElastic;
use gooey::animation::{AnimationTarget, Spawn};
use gooey::value::Dynamic;

let value = Dynamic::new(0);
let mut reader = value.create_reader();
value
    .transition_to(100)
    .over(Duration::from_millis(100))
    .with_easing(EaseInOutElastic)
    .launch();
drop(value);

while reader.block_until_updated() {
    println!("{}", reader.get());
}

assert_eq!(reader.get(), 100);

Modules

easings

Built-in Easing implementations.

Structs

Animation

Describes a change to a new value for a Dynamic over a specified Duration, using the Easing generic parameter to control how the value is interpolated.

AnimationHandle

A handle to a spawned animation. When dropped, the associated animation will be stopped.

BinaryLerp

A wrapper that implements LinearInterpolate such that the value switches after 50%.

BoxedAnimation

A Animate implementor that has been boxed as a trait object.

Chain

An animation combinator that runs animation A, then animation B.

Cycle

An animation that repeats another animation.

DynamicTransition

A pending transition for a Dynamic to a new value.

ImmediateLerp

A wrapper that implements LinearInterpolate such that the target value is immediately returned as long as percent is > 0.

OnCompleteAnimation

An animation wrapper that invokes a callback upon the animation completing.

RunningAnimation

A running Animation that changes a Dynamic over a specified Duration, using the Easing generic parameter to control how the value is interpolated.

RunningChain

A Chain that is currently animating.

TransitioningDynamic

A DynamicTransition that has begun its transition.

ZeroToOne

An f32 that is clamped between 0.0 and 1.0 and cannot be NaN or Infinity.

Enums

EasingFunction

An easing function for customizing animations.

Traits

Animate

A type that can animate.

AnimateTarget

The target of an Animate implementor.

AnimationTarget

A target for a timed Animation.

BoxAnimate

A type that can convert into Box<dyn Animate>.

Easing

Performs easing for value interpolation.

IntoAnimate

A type that can be converted into an animation.

LinearInterpolate

Performs a linear interpolation between two values.

PercentBetween

Calculates the ratio of one value against a minimum and maximum.

Spawn

An animation that can be spawned.

Derive Macros

LinearInterpolate

Derives LinerarInterpolate for structs and fieldless enums.