Module gooey::animation::easings

Built-in Easing implementations.

Structs

EaseInBack

An Easing function that eases in using a curve that backs away initially .

EaseInBounce

An Easing function that eases in using a curve that bounces progressively closer as it progresses .

EaseInCircular

An Easing function that eases in using a curve resembling the top-left arc of a circle .

EaseInCubic

An Easing function that eases in using a cubic (x^3) curve .

EaseInElastic

An Easing function that eases in using a curve that bounces around the start initially then quickly accelerates .

EaseInExponential

An Easing function that eases in using an expenential curve .

EaseInOutBack

An Easing function that eases in and out using a curve that backs away initially .

EaseInOutCircular

An Easing function that eases in and out using a curve resembling the top-left arc of a circle .

EaseInOutCubic

An Easing function that eases in and out using a cubic (x^3) curve .

EaseInOutElastic

An Easing function that eases in and out using a curve that bounces around the start initially then quickly accelerates .

EaseInOutExponential

An Easing function that eases in and out using an expenential curve .

EaseInOutQuadradic

An Easing function that eases in and out using a quadradic (x^2) curve .

EaseInOutQuartic

An Easing function that eases in and out using a quartic (x^4) curve .

EaseInOutQuintic

An Easing function that eases in and out using a quintic (x^5) curve .

EaseInOutSine

An Easing function that eases in and out using a sine wave .

EaseInQuadradic

An Easing function that eases in using a quadradic (x^2) curve .

EaseInQuartic

An Easing function that eases in using a quartic (x^4) curve .

EaseInQuintic

An Easing function that eases in using a quintic (x^5) curve .

EaseOutBack

An Easing function that eases out using a curve that backs away initially .

EaseOutBounce

An Easing function that eases out using a curve that bounces progressively closer as it progresses .

EaseOutCircular

An Easing function that eases out using a curve resembling the top-left arc of a circle .

EaseOutCubic

An Easing function that eases out using a cubic (x^3) curve .

EaseOutElastic

An Easing function that eases out using a curve that bounces around the start initially then quickly accelerates .

EaseOutExponential

An Easing function that eases out using an expenential curve .

EaseOutQuadradic

An Easing function that eases out using a quadradic (x^2) curve .

EaseOutQuartic

An Easing function that eases out using a quartic (x^4) curve .

EaseOutQuintic

An Easing function that eases out using a quintic (x^5) curve .

EaseOutSine

An Easing function that eases out using a sine wave .

Linear

An Easing function that produces a steady, linear transition.