Struct gooey::animation::ZeroToOne

pub struct ZeroToOne(_);

An f32 that is clamped between 0.0 and 1.0 and cannot be NaN or Infinity.

Because of these restrictions, this type implements Ord and Eq.

Implementations

impl ZeroToOne

pub const ONE: Self = _

The maximum value this type can contain.

pub const ZERO: Self = _

The minimum type this type can contain.

pub fn new(value: f32) -> Self

Returns a new instance after clamping value between +0.0 and 1.0.

Panics

This function panics if value is not a number.

pub fn into_f32(self) -> f32

Returns the contained floating point value.