Trait gooey::animation::IntoAnimate

pub trait IntoAnimate: Sized + Send + Sync {
    type Animate: Animate;

    // Required method
    fn into_animate(self) -> Self::Animate;

    // Provided methods
    fn and_then<Other: IntoAnimate>(self, other: Other) -> Chain<Self, Other> { ... }
    fn cycle(self) -> Cycle<Self>
       where Self: Clone { ... }
    fn repeat(self, times: usize) -> Cycle<Self>
       where Self: Clone { ... }
    fn on_complete<F>(self, on_complete: F) -> OnCompleteAnimation<Self>
       where F: FnMut() + Send + Sync + 'static { ... }
}

A type that can be converted into an animation.

Required Associated Types

type Animate: Animate

The running animation type.

Required Methods

fn into_animate(self) -> Self::Animate

Return this change as a running animation.

Provided Methods

fn and_then<Other: IntoAnimate>(self, other: Other) -> Chain<Self, Other>

Returns an combined animation that performs self and other in sequence.

fn cycle(self) -> Cycle<Self>where Self: Clone,

Returns an animation that repeats self indefinitely.

fn repeat(self, times: usize) -> Cycle<Self>where Self: Clone,

Returns an animation that repeats a number of times before completing.

fn on_complete<F>(self, on_complete: F) -> OnCompleteAnimation<Self>where F: FnMut() + Send + Sync + 'static,

Invokes on_complete after this animation finishes.