Trait gooey::animation::PercentBetween

pub trait PercentBetween {
    // Required method
    fn percent_between(&self, min: &Self, max: &Self) -> ZeroToOne;
}

Calculates the ratio of one value against a minimum and maximum.

Required Methods

fn percent_between(&self, min: &Self, max: &Self) -> ZeroToOne

Return the percentage that self is between min and max.