Trait gooey::animation::Spawn

pub trait Spawn {
    // Required method
    fn spawn(self) -> AnimationHandle;
}

An animation that can be spawned.

Required Methods

fn spawn(self) -> AnimationHandle

Spawns the animation, returning a handle that tracks the animation.

When the returned handle is dropped, the animation is stopped.