Struct gooey::context::EventContext

pub struct EventContext<'context, 'window> {
    pub widget: WidgetContext<'context, 'window>,
    pub kludgine: &'context mut Kludgine,
}

A context to an event function.

This type is a combination of a reference to the rendering library, [Kludgine], and a WidgetContext.

Fields
widget: WidgetContext<'context, 'window>

The context for the widget receiving the event.

kludgine: &'context mut Kludgine

The rendering library’s state.

This is useful for accessing the current scale or information needed to measure and layout text.

Implementations

impl<'context, 'window> EventContext<'context, 'window>

pub fn for_other<'child>( &'child mut self, widget: ManagedWidget ) -> EventContext<'child, 'window>

Returns a new EventContext with widget being referenced in the contained WidgetContext.

This function is used when one widget contains other widgets, and the parent widget needs to invoke events on a child widget. This is done by creating an EventContext pointing to the child and calling the appropriate function to invoke the event.

pub fn hit_test(&mut self, location: Point<Px>) -> bool

Invokes Widget::hit_test() on this context’s widget and returns the result.

pub fn mouse_down( &mut self, location: Point<Px>, device_id: DeviceId, button: MouseButton ) -> EventHandling

Invokes Widget::mouse_down() on this context’s widget and returns the result.

pub fn mouse_drag( &mut self, location: Point<Px>, device_id: DeviceId, button: MouseButton )

Invokes Widget::hit_test() on this context’s widget and returns the result.

pub fn mouse_up( &mut self, location: Option<Point<Px>>, device_id: DeviceId, button: MouseButton )

Invokes Widget::mouse_up() on this context’s widget and returns the result.

pub fn keyboard_input( &mut self, device_id: DeviceId, input: KeyEvent, is_synthetic: bool ) -> EventHandling

Invokes Widget::keyboard_input() on this context’s widget and returns the result.

pub fn ime(&mut self, ime: Ime) -> EventHandling

Invokes Widget::ime() on this context’s widget and returns the result.

pub fn mouse_wheel( &mut self, device_id: DeviceId, delta: MouseScrollDelta, phase: TouchPhase ) -> EventHandling

Invokes Widget::mouse_wheel() on this context’s widget and returns the result.