Struct gooey::context::GraphicsContext

pub struct GraphicsContext<'context, 'window, 'clip, 'gfx, 'pass> {
    pub widget: WidgetContext<'context, 'window>,
    pub gfx: Exclusive<'context, Graphics<'clip, 'gfx, 'pass>>,
}

A context to a function that is rendering a widget.

Fields
widget: WidgetContext<'context, 'window>

The context of the widget being rendered.

gfx: Exclusive<'context, Graphics<'clip, 'gfx, 'pass>>

The graphics context clipped and offset to the area of the widget being rendered. Drawing at 0,0 will draw at the top-left pixel of the laid-out widget region.

Implementations

impl<'context, 'window, 'clip, 'gfx, 'pass> GraphicsContext<'context, 'window, 'clip, 'gfx, 'pass>

pub fn borrowed(&mut self) -> GraphicsContext<'_, 'window, 'clip, 'gfx, 'pass>

Returns a new instance that borrows from self.

pub fn for_other<'child, Widget>( &'child mut self, widget: &Widget ) -> <Widget::Managed as MapManagedWidget<GraphicsContext<'child, 'window, 'child, 'gfx, 'pass>>>::Resultwhere Widget: ManageWidget, Widget::Managed: MapManagedWidget<GraphicsContext<'child, 'window, 'child, 'gfx, 'pass>>,

Returns a new GraphicsContext that allows invoking graphics functions for widget.

pub fn clipped_to( &mut self, clip: Rect<Px> ) -> GraphicsContext<'_, 'window, '_, 'gfx, 'pass>

Returns a new graphics context that renders to the clip rectangle.

pub fn fill(&mut self, color: Color)

Fills the background of this widget with color, honoring the current CornerRadius setting.

If the alpha channel of color is 0, this function does nothing.

pub fn stroke_outline<Unit>( &mut self, color: Color, options: StrokeOptions<Unit> )where Unit: ScreenScale<Px = Px, Lp = Lp, UPx = UPx> + Zero,

Strokes an outline around this widget’s contents.

pub fn draw_focus_ring(&mut self)

Renders the default focus ring for this widget.

pub fn apply_current_font_settings(&mut self)

Applies the current style settings for font family, text size, font style, and font weight.

pub fn redraw(&mut self)

Invokes Widget::redraw() on this context’s widget.

Panics

This function panics if the widget being drawn has no layout set (via LayoutContext::set_child_layout()).