Struct gooey::core::euclid::Rect

#[repr(C)]
pub struct Rect<T, U> {
    pub origin: Point2D<T, U>,
    pub size: Size2D<T, U>,
}

A 2d Rectangle optionally tagged with a unit.

Representation

Rect is represented by an origin point and a size.

See Box2D for a rectangle represented by two endpoints.

Empty rectangle

A rectangle is considered empty (see is_empty) if any of the following is true:

it’s area is empty,

it’s area is negative (size.x < 0 or size.y < 0),

it contains NaNs.

Fields
origin: Point2D<T, U>size: Size2D<T, U>
Implementations

impl<T, U> Rect<T, U>

pub const fn new(origin: Point2D<T, U>, size: Size2D<T, U>) -> Rect<T, U>

Constructor.

impl<T, U> Rect<T, U> where
T: Zero,

pub fn zero() -> Rect<T, U>

Constructor, setting all sides to zero.

pub fn from_size(size: Size2D<T, U>) -> Rect<T, U>

Creates a rect of the given size, at offset zero.

impl<T, U> Rect<T, U> where
T: Copy + Add<T, Output = T>,

pub fn min(&self) -> Point2D<T, U>

pub fn max(&self) -> Point2D<T, U>

pub fn max_x(&self) -> T

pub fn min_x(&self) -> T

pub fn max_y(&self) -> T

pub fn min_y(&self) -> T

pub fn width(&self) -> T

pub fn height(&self) -> T

pub fn x_range(&self) -> Range<T>

pub fn y_range(&self) -> Range<T>

#[must_use]
pub fn translate(&self, by: Vector2D<T, U>) -> Rect<T, U>

Returns the same rectangle, translated by a vector.

pub fn to_box2d(&self) -> Box2D<T, U>

impl<T, U> Rect<T, U> where
T: Copy + PartialOrd<T> + Add<T, Output = T>,

pub fn contains(&self, p: Point2D<T, U>) -> bool

Returns true if this rectangle contains the point. Points are considered in the rectangle if they are on the left or top edge, but outside if they are on the right or bottom edge.

pub fn intersects(&self, other: &Rect<T, U>) -> bool

impl<T, U> Rect<T, U> where
T: Copy + PartialOrd<T> + Add<T, Output = T> + Sub<T, Output = T>,

pub fn intersection(&self, other: &Rect<T, U>) -> Option<Rect<T, U>>

impl<T, U> Rect<T, U> where
T: Copy + Add<T, Output = T> + Sub<T, Output = T>,

#[must_use]
pub fn inflate(&self, width: T, height: T) -> Rect<T, U>

impl<T, U> Rect<T, U> where
T: Copy + Zero + PartialOrd<T> + Add<T, Output = T>,

pub fn contains_rect(&self, rect: &Rect<T, U>) -> bool

Returns true if this rectangle contains the interior of rect. Always returns true if rect is empty, and always returns false if rect is nonempty but this rectangle is empty.

impl<T, U> Rect<T, U> where
T: Copy + Zero + PartialOrd<T> + Add<T, Output = T> + Sub<T, Output = T>,

pub fn inner_rect(&self, offsets: SideOffsets2D<T, U>) -> Rect<T, U>

Calculate the size and position of an inner rectangle.

Subtracts the side offsets from all sides. The horizontal and vertical offsets must not be larger than the original side length. This method assumes y oriented downward.

impl<T, U> Rect<T, U> where
T: Copy + Add<T, Output = T> + Sub<T, Output = T>,

pub fn outer_rect(&self, offsets: SideOffsets2D<T, U>) -> Rect<T, U>

Calculate the size and position of an outer rectangle.

Add the offsets to all sides. The expanded rectangle is returned. This method assumes y oriented downward.

impl<T, U> Rect<T, U> where
T: Copy + Zero + PartialOrd<T> + Sub<T, Output = T>,

pub fn from_points<I>(points: I) -> Rect<T, U> where
I: IntoIterator,
<I as IntoIterator>::Item: Borrow<Point2D<T, U>>,

Returns the smallest rectangle defined by the top/bottom/left/right-most points provided as parameter.

Note: This function has a behavior that can be surprising because the right-most and bottom-most points are exactly on the edge of the rectangle while the contains function is has exclusive semantic on these edges. This means that the right-most and bottom-most points provided to from_points will count as not contained by the rect. This behavior may change in the future.

impl<T, U> Rect<T, U> where
T: One + Add<T, Output = T> + Sub<T, Output = T> + Mul<T, Output = T> + Copy,

pub fn lerp(&self, other: Rect<T, U>, t: T) -> Rect<T, U>

Linearly interpolate between this rectangle and another rectangle.

impl<T, U> Rect<T, U> where
T: Copy + One + Add<T, Output = T> + Div<T, Output = T>,

pub fn center(&self) -> Point2D<T, U>

impl<T, U> Rect<T, U> where
T: Copy + PartialOrd<T> + Add<T, Output = T> + Sub<T, Output = T> + Zero,

pub fn union(&self, other: &Rect<T, U>) -> Rect<T, U>

impl<T, U> Rect<T, U>

pub fn scale<S>(&self, x: S, y: S) -> Rect<T, U> where
T: Copy + Mul<S, Output = T>,
S: Copy,

impl<T, U> Rect<T, U> where
T: Copy + Mul<T, Output = T>,

pub fn area(&self) -> T

impl<T, U> Rect<T, U> where
T: Copy + Zero + PartialOrd<T>,

pub fn is_empty(&self) -> bool

impl<T, U> Rect<T, U> where
T: Copy + Zero + PartialOrd<T>,

pub fn to_non_empty(&self) -> Option<Rect<T, U>>

impl<T, U> Rect<T, U> where
T: Copy,

pub fn to_untyped(&self) -> Rect<T, UnknownUnit>

Drop the units, preserving only the numeric value.

pub fn from_untyped(r: &Rect<T, UnknownUnit>) -> Rect<T, U>

Tag a unitless value with units.

pub fn cast_unit<V>(&self) -> Rect<T, V>

Cast the unit

impl<T, U> Rect<T, U> where
T: NumCast + Copy,

pub fn cast<NewT>(&self) -> Rect<NewT, U> where
NewT: NumCast,

Cast from one numeric representation to another, preserving the units.

When casting from floating point to integer coordinates, the decimals are truncated as one would expect from a simple cast, but this behavior does not always make sense geometrically. Consider using round(), round_in or round_out() before casting.

pub fn try_cast<NewT>(&self) -> Option<Rect<NewT, U>> where
NewT: NumCast,

Fallible cast from one numeric representation to another, preserving the units.

When casting from floating point to integer coordinates, the decimals are truncated as one would expect from a simple cast, but this behavior does not always make sense geometrically. Consider using round(), round_in or round_out() before casting.

pub fn to_f32(&self) -> Rect<f32, U>

Cast into an f32 rectangle.

pub fn to_f64(&self) -> Rect<f64, U>

Cast into an f64 rectangle.

pub fn to_usize(&self) -> Rect<usize, U>

Cast into an usize rectangle, truncating decimals if any.

When casting from floating point rectangles, it is worth considering whether to round(), round_in() or round_out() before the cast in order to obtain the desired conversion behavior.

pub fn to_u32(&self) -> Rect<u32, U>

Cast into an u32 rectangle, truncating decimals if any.

When casting from floating point rectangles, it is worth considering whether to round(), round_in() or round_out() before the cast in order to obtain the desired conversion behavior.

pub fn to_u64(&self) -> Rect<u64, U>

Cast into an u64 rectangle, truncating decimals if any.

When casting from floating point rectangles, it is worth considering whether to round(), round_in() or round_out() before the cast in order to obtain the desired conversion behavior.

pub fn to_i32(&self) -> Rect<i32, U>

Cast into an i32 rectangle, truncating decimals if any.

When casting from floating point rectangles, it is worth considering whether to round(), round_in() or round_out() before the cast in order to obtain the desired conversion behavior.

pub fn to_i64(&self) -> Rect<i64, U>

Cast into an i64 rectangle, truncating decimals if any.

When casting from floating point rectangles, it is worth considering whether to round(), round_in() or round_out() before the cast in order to obtain the desired conversion behavior.

impl<T, U> Rect<T, U> where
T: Float,

pub fn is_finite(self) -> bool

Returns true if all members are finite.

impl<T, U> Rect<T, U> where
T: Ceil + Round + Add<T, Output = T> + Sub<T, Output = T> + Floor,

#[must_use]
pub fn round(&self) -> Rect<T, U>

Return a rectangle with edges rounded to integer coordinates, such that the returned rectangle has the same set of pixel centers as the original one. Edges at offset 0.5 round up. Suitable for most places where integral device coordinates are needed, but note that any translation should be applied first to avoid pixel rounding errors. Note that this is not rounding to nearest integer if the values are negative. They are always rounding as floor(n + 0.5).

Usage notes

Note, that when using with floating-point T types that method can significantly loose precision for large values, so if you need to call this method very often it is better to use Box2D.

#[must_use]
pub fn round_in(&self) -> Rect<T, U>

Return a rectangle with edges rounded to integer coordinates, such that the original rectangle contains the resulting rectangle.

Usage notes

Note, that when using with floating-point T types that method can significantly loose precision for large values, so if you need to call this method very often it is better to use Box2D.

#[must_use]
pub fn round_out(&self) -> Rect<T, U>

Return a rectangle with edges rounded to integer coordinates, such that the original rectangle is contained in the resulting rectangle.

Usage notes

Note, that when using with floating-point T types that method can significantly loose precision for large values, so if you need to call this method very often it is better to use Box2D.