Struct gooey::core::palette::Hsl

#[repr(C)]
pub struct Hsl<S = Srgb, T = f32> where
    T: Component + Float,
    S: RgbSpace,  {
    pub hue: RgbHue<T>,
    pub saturation: T,
    pub lightness: T,
    pub space: PhantomData<S>,
}

Linear HSL color space.

The HSL color space can be seen as a cylindrical version of RGB, where the hue is the angle around the color cylinder, the saturation is the distance from the center, and the lightness is the height from the bottom. Its composition makes it especially good for operations like changing green to red, making a color more gray, or making it darker.

See HSV for a very similar color space, with brightness instead of lightness.

Fields
hue: RgbHue<T>

The hue of the color, in degrees. Decides if it’s red, blue, purple, etc.

saturation: T

The colorfulness of the color. 0.0 gives gray scale colors and 1.0 will give absolutely clear colors.

lightness: T

Decides how light the color will look. 0.0 will be black, 0.5 will give a clear color, and 1.0 will give white.

space: PhantomData<S>

The white point and RGB primaries this color is adapted to. The default is the sRGB standard.

Implementations

impl<T> Hsl<Srgb, T> where
T: Component + Float,

pub fn new<H>(hue: H, saturation: T, lightness: T) -> Hsl<Srgb, T> where
H: Into<RgbHue<T>>,

HSL for linear sRGB.

impl<S, T> Hsl<S, T> where
T: Component + Float,
S: RgbSpace,

pub fn with_wp<H>(hue: H, saturation: T, lightness: T) -> Hsl<S, T> where
H: Into<RgbHue<T>>,

Linear HSL.

pub fn into_components(self) -> (RgbHue<T>, T, T)

Convert to a (hue, saturation, lightness) tuple.

pub fn from_components<H>((H, T, T)) -> Hsl<S, T> where
H: Into<RgbHue<T>>,

Convert from a (hue, saturation, lightness) tuple.