Trait gooey::core::palette::Clamp

pub trait Clamp {
    fn is_within_bounds(&self) -> bool;

    fn clamp(&self) -> Self;

    fn clamp_self(&mut self);
}

A trait for clamping and checking if colors are within their ranges.

Required methods

fn is_within_bounds(&self) -> bool

Check if the color’s components are within the expected clamped range bounds.

fn clamp(&self) -> Self

Return a new color where the components have been clamped to the nearest valid values.

fn clamp_self(&mut self)

Clamp the color’s components to the nearest valid values.