Trait gooey::core::palette::Hue

pub trait Hue: GetHue {
    fn with_hue<H>(&self, hue: H) -> Self
    where
        H: Into<Self::Hue>;

    fn shift_hue<H>(&self, amount: H) -> Self
    where
        H: Into<Self::Hue>;
}

A trait for colors where the hue can be manipulated without conversion.

Required methods

fn with_hue<H>(&self, hue: H) -> Self where
H: Into<Self::Hue>,

Return a new copy of self, but with a specific hue.

fn shift_hue<H>(&self, amount: H) -> Self where
H: Into<Self::Hue>,

Return a new copy of self, but with the hue shifted by amount.