Trait gooey::core::renderer::Renderer

pub trait Renderer: 'static + Debug + Send + Sync {
    pub fn size(&self) -> Size2D<f32, Points>;

    pub fn clip_to(&self, bounds: Rect<f32, Points>) -> Self;

    pub fn clip_bounds(&self) -> Rect<f32, Points>;

    pub fn scale(&self) -> Scale<f32, Points, Pixels>;

    pub fn render_text<F>(
        &self, 
        text: &str, 
        baseline_origin: Point2D<f32, Points>, 
        style: &Style
    )
    where
        F: FallbackComponent<Value = ColorPair>;

#[must_use]    pub fn measure_text(&self, text: &str, style: &Style) -> TextMetrics<Points>;

    pub fn stroke_rect(&self, rect: &Rect<f32, Points>, style: &Style);

    pub fn fill_rect<F>(&self, rect: &Rect<f32, Points>, style: &Style)
    where
        F: FallbackComponent<Value = ColorPair>;

    pub fn stroke_line(
        &self, 
        point_a: Point2D<f32, Points>, 
        point_b: Point2D<f32, Points>, 
        style: &Style
    );

    pub fn bounds(&self) -> Rect<f32, Points> { ... }
}

Implements drawing APIs.

Required methods

pub fn size(&self) -> Size2D<f32, Points>

The size of the area being drawn.

pub fn clip_to(&self, bounds: Rect<f32, Points>) -> Self

Returns a new renderer instance with the state such that each operation executes as if the origin is bounds.origin. The returned instance’s size() should equal bounds.size.

pub fn clip_bounds(&self) -> Rect<f32, Points>

A Rect representing the area being drawn. This rect should be offset relative to the origin of the renderer.

pub fn scale(&self) -> Scale<f32, Points, Pixels>

The scale when converting between Points and Pixels.

pub fn render_text<F>(
&self,
text: &str,
baseline_origin: Point2D<f32, Points>,
style: &Style
) where
F: FallbackComponent<Value = ColorPair>,

Renders text at baseline_origin with options.

#[must_use]pub fn measure_text(&self, text: &str, style: &Style) -> TextMetrics<Points>

Measures text using options.

pub fn stroke_rect(&self, rect: &Rect<f32, Points>, style: &Style)

Strokes the outline of rect using options.

pub fn fill_rect<F>(&self, rect: &Rect<f32, Points>, style: &Style) where
F: FallbackComponent<Value = ColorPair>,

Fills rect using color.

pub fn stroke_line(
&self,
point_a: Point2D<f32, Points>,
point_b: Point2D<f32, Points>,
style: &Style
)

Draws a line between point_a and point_b using options.

Loading content...
Provided methods

pub fn bounds(&self) -> Rect<f32, Points>

A Rect representing the area being drawn. Due to how rendering works, the origin is always zero.

Loading content...