Module gooey::core::stylecs

Modules

palette
A library that makes linear color calculations and conversion easy and accessible for anyone. It uses the type system to enforce correctness and to avoid mistakes, such as mixing incompatible color types.

Structs

BackgroundColor

ColorPair

ComponentCollection

FontFamily

FontSize

ForegroundColor

Pixels

Points

Style

StyleSheet

Enums

Alignment

FontStyle

GenericStyle

SystemTheme

VerticalAlignment

Weight

Traits

AnyStyleComponent

StyleComponent

UnscaledStyleComponent