Struct gooey::core::styles::Color

pub struct Color {
    pub red: f32,
    pub green: f32,
    pub blue: f32,
    pub alpha: f32,
}

An Srgba color.

Fields
red: f32

The red channel. Valid values are from 0.0 to 1.0 inclusive.

green: f32

The green channel. Valid values are from 0.0 to 1.0 inclusive.

blue: f32

The blue channel. Valid values are from 0.0 to 1.0 inclusive.

alpha: f32

The alpha channel. Valid values are from 0.0 to 1.0 inclusive, where 1.0 is fully opaque.

Implementations

impl Color

pub const fn new(red: f32, green: f32, blue: f32, alpha: f32) -> Color

Creates a new color with SRGBA components red, green, blue, and alpha ranging from 0.0-1.0.

pub fn new_u8(red: u8, green: u8, blue: u8, alpha: u8) -> Color

Creates a new color with SRGBA components red, green, blue, and alpha ranging from 0-255.

pub fn as_css_string(&self) -> String

Formats the color for CSS.

pub const fn gray(lightness: f32) -> Color

Returns a gray color with lightness in the red, green, and blue channels and an alpha of 1.0.

pub const fn with_alpha(&self, alpha: f32) -> Color

Returns a new color with the same red, green, and blue values, but replacing the current alpha with the parameter.

pub fn lighten(self, factor: f32) -> Color

Returns this color after lightening it by factor. Uses the Laba colorspace for luminance calcluations.

pub fn darken(self, factor: f32) -> Color

Returns this color after darkening it by factor. Uses the Laba colorspace for luminance calcluations.

impl Color

pub const ALICEBLUE: Color

Equivalent to the CSS color keywords of the same name.

pub const ANTIQUEWHITE: Color

Equivalent to the CSS color keywords of the same name.

pub const AQUA: Color

Equivalent to the CSS color keywords of the same name.

pub const AQUAMARINE: Color

Equivalent to the CSS color keywords of the same name.

pub const AZURE: Color

Equivalent to the CSS color keywords of the same name.

pub const BEIGE: Color

Equivalent to the CSS color keywords of the same name.

pub const BISQUE: Color

Equivalent to the CSS color keywords of the same name.

pub const BLACK: Color

Equivalent to the CSS color keywords of the same name.

pub const BLANCHEDALMOND: Color

Equivalent to the CSS color keywords of the same name.

pub const BLUE: Color

Equivalent to the CSS color keywords of the same name.

pub const BLUEVIOLET: Color

Equivalent to the CSS color keywords of the same name.

pub const BROWN: Color

Equivalent to the CSS color keywords of the same name.

pub const BURLYWOOD: Color

Equivalent to the CSS color keywords of the same name.

pub const CADETBLUE: Color

Equivalent to the CSS color keywords of the same name.

pub const CHARTREUSE: Color

Equivalent to the CSS color keywords of the same name.

pub const CHOCOLATE: Color

Equivalent to the CSS color keywords of the same name.

pub const CLEAR_BLACK: Color

Equivalent to the CSS color keywords of the same name.

pub const CLEAR_WHITE: Color

Equivalent to the CSS color keywords of the same name.

pub const CORAL: Color

Equivalent to the CSS color keywords of the same name.

pub const CORNFLOWERBLUE: Color

Equivalent to the CSS color keywords of the same name.

pub const CORNSILK: Color

Equivalent to the CSS color keywords of the same name.

pub const CRIMSON: Color

Equivalent to the CSS color keywords of the same name.

pub const CYAN: Color

Equivalent to the CSS color keywords of the same name.

pub const DARKBLUE: Color

Equivalent to the CSS color keywords of the same name.

pub const DARKCYAN: Color

Equivalent to the CSS color keywords of the same name.

pub const DARKGOLDENROD: Color

Equivalent to the CSS color keywords of the same name.

pub const DARKGRAY: Color

Equivalent to the CSS color keywords of the same name.

pub const DARKGREEN: Color

Equivalent to the CSS color keywords of the same name.

pub const DARKGREY: Color

Equivalent to the CSS color keywords of the same name.

pub const DARKKHAKI: Color

Equivalent to the CSS color keywords of the same name.

pub const DARKMAGENTA: Color

Equivalent to the CSS color keywords of the same name.

pub const DARKOLIVEGREEN: Color

Equivalent to the CSS color keywords of the same name.

pub const DARKORANGE: Color

Equivalent to the CSS color keywords of the same name.

pub const DARKORCHID: Color

Equivalent to the CSS color keywords of the same name.

pub const DARKRED: Color

Equivalent to the CSS color keywords of the same name.

pub const DARKSALMON: Color

Equivalent to the CSS color keywords of the same name.

pub const DARKSEAGREEN: Color

Equivalent to the CSS color keywords of the same name.

pub const DARKSLATEBLUE: Color

Equivalent to the CSS color keywords of the same name.

pub const DARKSLATEGRAY: Color

Equivalent to the CSS color keywords of the same name.

pub const DARKSLATEGREY: Color

Equivalent to the CSS color keywords of the same name.

pub const DARKTURQUOISE: Color

Equivalent to the CSS color keywords of the same name.

pub const DARKVIOLET: Color

Equivalent to the CSS color keywords of the same name.

pub const DEEPPINK: Color

Equivalent to the CSS color keywords of the same name.

pub const DEEPSKYBLUE: Color

Equivalent to the CSS color keywords of the same name.

pub const DIMGRAY: Color

Equivalent to the CSS color keywords of the same name.

pub const DIMGREY: Color

Equivalent to the CSS color keywords of the same name.

pub const DODGERBLUE: Color

Equivalent to the CSS color keywords of the same name.

pub const FIREBRICK: Color

Equivalent to the CSS color keywords of the same name.

pub const FLORALWHITE: Color

Equivalent to the CSS color keywords of the same name.

pub const FORESTGREEN: Color

Equivalent to the CSS color keywords of the same name.

pub const FUCHSIA: Color

Equivalent to the CSS color keywords of the same name.

pub const GAINSBORO: Color

Equivalent to the CSS color keywords of the same name.

pub const GHOSTWHITE: Color

Equivalent to the CSS color keywords of the same name.

pub const GOLD: Color

Equivalent to the CSS color keywords of the same name.

pub const GOLDENROD: Color

Equivalent to the CSS color keywords of the same name.

pub const GRAY: Color

Equivalent to the CSS color keywords of the same name.

pub const GREEN: Color

Equivalent to the CSS color keywords of the same name.

pub const GREENYELLOW: Color

Equivalent to the CSS color keywords of the same name.

pub const GREY: Color

Equivalent to the CSS color keywords of the same name.

pub const HONEYDEW: Color

Equivalent to the CSS color keywords of the same name.

pub const HOTPINK: Color

Equivalent to the CSS color keywords of the same name.

pub const INDIANRED: Color

Equivalent to the CSS color keywords of the same name.

pub const INDIGO: Color

Equivalent to the CSS color keywords of the same name.

pub const IVORY: Color

Equivalent to the CSS color keywords of the same name.

pub const KHAKI: Color

Equivalent to the CSS color keywords of the same name.

pub const LAVENDER: Color

Equivalent to the CSS color keywords of the same name.

pub const LAVENDERBLUSH: Color

Equivalent to the CSS color keywords of the same name.

pub const LAWNGREEN: Color

Equivalent to the CSS color keywords of the same name.

pub const LEMONCHIFFON: Color

Equivalent to the CSS color keywords of the same name.

pub const LIGHTBLUE: Color

Equivalent to the CSS color keywords of the same name.

pub const LIGHTCORAL: Color

Equivalent to the CSS color keywords of the same name.

pub const LIGHTCYAN: Color

Equivalent to the CSS color keywords of the same name.

pub const LIGHTGOLDENRODYELLOW: Color

Equivalent to the CSS color keywords of the same name.

pub const LIGHTGRAY: Color

Equivalent to the CSS color keywords of the same name.

pub const LIGHTGREEN: Color

Equivalent to the CSS color keywords of the same name.

pub const LIGHTGREY: Color

Equivalent to the CSS color keywords of the same name.

pub const LIGHTPINK: Color

Equivalent to the CSS color keywords of the same name.

pub const LIGHTSALMON: Color

Equivalent to the CSS color keywords of the same name.

pub const LIGHTSEAGREEN: Color

Equivalent to the CSS color keywords of the same name.

pub const LIGHTSKYBLUE: Color

Equivalent to the CSS color keywords of the same name.

pub const LIGHTSLATEGRAY: Color

Equivalent to the CSS color keywords of the same name.

pub const LIGHTSLATEGREY: Color

Equivalent to the CSS color keywords of the same name.

pub const LIGHTSTEELBLUE: Color

Equivalent to the CSS color keywords of the same name.

pub const LIGHTYELLOW: Color

Equivalent to the CSS color keywords of the same name.

pub const LIME: Color

Equivalent to the CSS color keywords of the same name.

pub const LIMEGREEN: Color

Equivalent to the CSS color keywords of the same name.

pub const LINEN: Color

Equivalent to the CSS color keywords of the same name.

pub const MAGENTA: Color

Equivalent to the CSS color keywords of the same name.

pub const MAROON: Color

Equivalent to the CSS color keywords of the same name.

pub const MEDIUMAQUAMARINE: Color

Equivalent to the CSS color keywords of the same name.

pub const MEDIUMBLUE: Color

Equivalent to the CSS color keywords of the same name.

pub const MEDIUMORCHID: Color

Equivalent to the CSS color keywords of the same name.

pub const MEDIUMPURPLE: Color

Equivalent to the CSS color keywords of the same name.

pub const MEDIUMSEAGREEN: Color

Equivalent to the CSS color keywords of the same name.

pub const MEDIUMSLATEBLUE: Color

Equivalent to the CSS color keywords of the same name.

pub const MEDIUMSPRINGGREEN: Color

Equivalent to the CSS color keywords of the same name.

pub const MEDIUMTURQUOISE: Color

Equivalent to the CSS color keywords of the same name.

pub const MEDIUMVIOLETRED: Color

Equivalent to the CSS color keywords of the same name.

pub const MIDNIGHTBLUE: Color

Equivalent to the CSS color keywords of the same name.

pub const MINTCREAM: Color

Equivalent to the CSS color keywords of the same name.

pub const MISTYROSE: Color

Equivalent to the CSS color keywords of the same name.

pub const MOCCASIN: Color

Equivalent to the CSS color keywords of the same name.

pub const NAVAJOWHITE: Color

Equivalent to the CSS color keywords of the same name.

pub const NAVY: Color

Equivalent to the CSS color keywords of the same name.

pub const OLDLACE: Color

Equivalent to the CSS color keywords of the same name.

pub const OLIVE: Color

Equivalent to the CSS color keywords of the same name.

pub const OLIVEDRAB: Color

Equivalent to the CSS color keywords of the same name.

pub const ORANGE: Color

Equivalent to the CSS color keywords of the same name.

pub const ORANGERED: Color

Equivalent to the CSS color keywords of the same name.

pub const ORCHID: Color

Equivalent to the CSS color keywords of the same name.

pub const PALEGOLDENROD: Color

Equivalent to the CSS color keywords of the same name.

pub const PALEGREEN: Color

Equivalent to the CSS color keywords of the same name.

pub const PALETURQUOISE: Color

Equivalent to the CSS color keywords of the same name.

pub const PALEVIOLETRED: Color

Equivalent to the CSS color keywords of the same name.

pub const PAPAYAWHIP: Color

Equivalent to the CSS color keywords of the same name.

pub const PEACHPUFF: Color

Equivalent to the CSS color keywords of the same name.

pub const PERU: Color

Equivalent to the CSS color keywords of the same name.

pub const PINK: Color

Equivalent to the CSS color keywords of the same name.

pub const PLUM: Color

Equivalent to the CSS color keywords of the same name.

pub const POWDERBLUE: Color

Equivalent to the CSS color keywords of the same name.

pub const PURPLE: Color

Equivalent to the CSS color keywords of the same name.

pub const REBECCAPURPLE: Color

Equivalent to the CSS color keywords of the same name.

pub const RED: Color

Equivalent to the CSS color keywords of the same name.

pub const ROSYBROWN: Color

Equivalent to the CSS color keywords of the same name.

pub const ROYALBLUE: Color

Equivalent to the CSS color keywords of the same name.

pub const SADDLEBROWN: Color

Equivalent to the CSS color keywords of the same name.

pub const SALMON: Color

Equivalent to the CSS color keywords of the same name.

pub const SANDYBROWN: Color

Equivalent to the CSS color keywords of the same name.

pub const SEAGREEN: Color

Equivalent to the CSS color keywords of the same name.

pub const SEASHELL: Color

Equivalent to the CSS color keywords of the same name.

pub const SIENNA: Color

Equivalent to the CSS color keywords of the same name.

pub const SILVER: Color

Equivalent to the CSS color keywords of the same name.

pub const SKYBLUE: Color

Equivalent to the CSS color keywords of the same name.

pub const SLATEBLUE: Color

Equivalent to the CSS color keywords of the same name.

pub const SLATEGRAY: Color

Equivalent to the CSS color keywords of the same name.

pub const SLATEGREY: Color

Equivalent to the CSS color keywords of the same name.

pub const SNOW: Color

Equivalent to the CSS color keywords of the same name.

pub const SPRINGGREEN: Color

Equivalent to the CSS color keywords of the same name.

pub const STEELBLUE: Color

Equivalent to the CSS color keywords of the same name.

pub const TAN: Color

Equivalent to the CSS color keywords of the same name.

pub const TEAL: Color

Equivalent to the CSS color keywords of the same name.

pub const THISTLE: Color

Equivalent to the CSS color keywords of the same name.

pub const TOMATO: Color

Equivalent to the CSS color keywords of the same name.

pub const TURQUOISE: Color

Equivalent to the CSS color keywords of the same name.

pub const VIOLET: Color

Equivalent to the CSS color keywords of the same name.

pub const WHEAT: Color

Equivalent to the CSS color keywords of the same name.

pub const WHITE: Color

Equivalent to the CSS color keywords of the same name.

pub const WHITESMOKE: Color

Equivalent to the CSS color keywords of the same name.

pub const YELLOW: Color

Equivalent to the CSS color keywords of the same name.

pub const YELLOWGREEN: Color

Equivalent to the CSS color keywords of the same name.