Trait gooey::core::AnyFrontend

pub trait AnyFrontend: AnySendSync {
    fn cloned(&self) -> Box<dyn AnyFrontend + 'static, Global>;
    fn storage(&self) -> &WidgetStorage;
    fn window(&self) -> Option<&(dyn Window + 'static)>;
    fn theme(&self) -> SystemTheme;
    fn set_widget_has_messages(&self, widget: WidgetId);
    fn enter_managed_code(&self) -> ManagedCodeGuard;
    fn load_image(
        &self, 
        asset: &Image, 
        completed: Callback<Image, ()>, 
        error: Callback<String, ()>
    );
    fn asset_url(&self, asset: &Asset) -> Option<Url>;
    fn schedule_timer(
        &self, 
        callback: Callback<(), ()>, 
        duration: Duration, 
        repeating: bool
    ) -> Timer;
    fn localize(
        &self, 
        key: &str, 
        parameters: Option<LocalizationParameters<'a>>
    ) -> String;
    fn open(&self, window: Box<dyn AnyWindowBuilder + 'static, Global>) -> bool;

    fn app(&self) -> &AppContext { ... }
}

An interface for Frontend that doesn’t requier knowledge of associated types.

Required Methods

fn cloned(&self) -> Box<dyn AnyFrontend + 'static, Global>

Clones the frontend, returning the clone in a box.

fn storage(&self) -> &WidgetStorage

Returns the widget storage.

fn window(&self) -> Option<&(dyn Window + 'static)>

Returns the window for this frontend instance.

fn theme(&self) -> SystemTheme

Returns the current system theme.

fn set_widget_has_messages(&self, widget: WidgetId)

Notifies the frontend that a widget has messages. Frontends should ensure that process_widget_messages is called at some point after this method is called.

fn enter_managed_code(&self) -> ManagedCodeGuard

Marks that managed code is being executed. Can be nested. Automatically exited when the returned guard is dropped.

fn load_image(
&self,
asset: &Image,
completed: Callback<Image, ()>,
error: Callback<String, ()>
)

Loads an image asynchronously, executing completed when loaded.

fn asset_url(&self, asset: &Asset) -> Option<Url>

Returns the full Url for the asset, if available.

fn schedule_timer(
&self,
callback: Callback<(), ()>,
duration: Duration,
repeating: bool
) -> Timer

Schedules a timer that invokes callback after duration, and repeats if repeating is true.

fn localize(
&self,
key: &str,
parameters: Option<LocalizationParameters<'a>>
) -> String

Localizes key with parameters.

fn open(&self, window: Box<dyn AnyWindowBuilder + 'static, Global>) -> bool

Opens a window. Returns false if unable to open the window.

Provided Methods

fn app(&self) -> &AppContext

Returns the current application context.