Trait gooey::core::KeyedStorage

pub trait KeyedStorage<K>: 'static + Debug + Send + Sync where
    K: Key, {
    fn register<W>(
        &mut self, 
        key: impl Into<Option<K>>, 
        styled_widget: StyledWidget<W>
    ) -> WidgetRegistration
    where
        W: Widget + AnyWidget;

    fn storage(&self) -> &WidgetStorage;

    fn related_storage(
        &self
    ) -> Option<Box<dyn RelatedStorage<K> + 'static, Global>>;
}

A type that registers widgets with an associated key.

Required methods

fn register<W>(
&mut self,
key: impl Into<Option<K>>,
styled_widget: StyledWidget<W>
) -> WidgetRegistration where
W: Widget + AnyWidget,

Register styled_widget with key.

fn storage(&self) -> &WidgetStorage

Returns the underlying widget storage.

fn related_storage(
&self
) -> Option<Box<dyn RelatedStorage<K> + 'static, Global>>

If this storage is representing a component, this returns a weak registration that can be used to communicate with it.