Trait gooey::core::RelatedStorage

pub trait RelatedStorage<K>: 'static + Debug + Send + Sync where
    K: Key, {
    pub fn widget(&self) -> WeakWidgetRegistration;

    pub fn remove(&self, key: &K);

    pub fn register(&self, key: K, widget: &WidgetRegistration);
}

Related storage enables a widget to communicate in a limited way about widgets being inserted or removed.

Required methods

pub fn widget(&self) -> WeakWidgetRegistration

Returns the registration of the widget that this is from.

pub fn remove(&self, key: &K)

Removes the widget with key from this storage.

pub fn register(&self, key: K, widget: &WidgetRegistration)

Registers widget with key.

Loading content...