Trait gooey::core::Window

pub trait Window: 'static + Debug + Send + Sync {
    fn set_title(&self, title: &str);

    fn inner_size(&self) -> Size<u32, Pixels>;

    fn set_inner_size(&self, new_size: Size<u32, Pixels>);

    fn set_outer_position(&self, new_position: Point<i32, Pixels>);

    fn inner_position(&self) -> Point<i32, Pixels>;

    fn set_always_on_top(&self, always: bool);

    fn maximized(&self) -> bool;

    fn set_maximized(&self, maximized: bool);

    fn set_minimized(&self, minimized: bool);

    fn close(&self);

    fn outer_position(&self) -> Point<i32, Pixels> { ... }
}

Represents a window.

Required methods

fn set_title(&self, title: &str)

Sets the window title.

fn inner_size(&self) -> Size<u32, Pixels>

Returns the size of the content area of the window.

fn set_inner_size(&self, new_size: Size<u32, Pixels>)

Attempts to resize the window to new_size. This may not work on all platforms.

fn set_outer_position(&self, new_position: Point<i32, Pixels>)

Sets the outer position of the window. This may not work on all platforms.

fn inner_position(&self) -> Point<i32, Pixels>

Returns the position of the top-left of the content area in screen coordinates.

fn set_always_on_top(&self, always: bool)

Sets whether the window should always be on top of other windows.

fn maximized(&self) -> bool

Returns true if the window is maximized.

fn set_maximized(&self, maximized: bool)

Sets whether the window should be maximized.

fn set_minimized(&self, minimized: bool)

Sets whether the window should be minimized.

fn close(&self)

Closes the window.

Provided methods

fn outer_position(&self) -> Point<i32, Pixels>

Returns the position on the screen of the window’s top-left corner. On platforms where this is unsupported, innner_position() is returned.