Function gooey::main

pub fn main<W: Widget, C: Fn(&WidgetStorage) -> StyledWidget<W>>(initializer: C)

Runs a Kludgine-based App with the root widget from initializer. All widgets from gooey::widget will be usable. If you wish to use other widgets, use browser_main_with and provide the transmogrifiers for the widgets you wish to use.