Struct gooey::frontends::renderers::kludgine::kludgine::core::math::ExtentsRect

pub struct ExtentsRect<T, Unit> {
    pub origin: Point<T, Unit>,
    pub extent: Point<T, Unit>,
}

A rectangle that uses two Points for representation.

Fields
origin: Point<T, Unit>

The origin of the rectangle.

extent: Point<T, Unit>

The non-origin point of the rectangle.

Implementations

impl<T, Unit> ExtentsRect<T, Unit>

pub fn new(
origin: Point<T, Unit>,
extent: Point<T, Unit>
) -> ExtentsRect<T, Unit>

Returns a new rectangle using origin and extent.

impl<T, Unit> ExtentsRect<T, Unit> where
T: NumCast + Copy,

pub fn try_cast<NewT>(&self) -> Option<ExtentsRect<NewT, Unit>> where
NewT: NumCast + Copy,

Attempts to cast T to NewT. If unsuccessful, None is returned.

pub fn cast<NewT>(&self) -> ExtentsRect<NewT, Unit> where
NewT: NumCast + Copy,

Casts T to NewT.

Panics

Panics if casting fails.

impl<T, Unit> ExtentsRect<T, Unit> where
T: Copy,

pub fn cast_unit<NewUnit>(&self) -> ExtentsRect<T, NewUnit>

Returns this value with the new unit. Does not affect the underlying value.

impl<T, Unit> ExtentsRect<T, Unit> where
T: Ceil + Floor,

pub fn round_out(self) -> ExtentsRect<T, Unit>

Returns a new rectangle that rounds the origin down using floor and rounds the extent out using ceil.

pub fn round_in(self) -> ExtentsRect<T, Unit>

Returns a new rectangle that rounds the origin up using ceil and rounds the extent in using floor.