Module gooey::frontends::renderers::kludgine::kludgine::easygpu::prelude

Modules

euclid
A collection of strongly typed math tools for computer graphics with an inclination towards 2d graphics and layout.

wgpu
A cross-platform graphics and compute library based on WebGPU.

Structs

Bgra8
A BGRA color with 8-bit channels, used when dealing with framebuffers.

Binding

BindingGroup
A group of bindings.

BindingGroupLayout
The layout of a BindingGroup.

Blending

DepthBuffer
Z-Depth buffer.

Device

Frame

Framebuffer
Off-screen framebuffer. Can be used as a render target in render passes.

IndexBuffer

Pipeline

PipelineCore

PipelineDescription

PipelineLayout

Renderer

Rgba

Rgba8

Sampler

ScreenSpace

Set

Shader

ShaderStage
Describes the shader stages that a binding will be visible from.

SwapChain
A handle to a swap chain.

SwapChainTexture

Texture

UniformBuffer
A uniform buffer that can be bound in a ‘BindingGroup’.

VertexBuffer

VertexLayout
Describes a ‘VertexBuffer’ layout.

WorldSpace

Enums

BindingType
A binding type.

BlendFactor

BlendOp

Error

Op

PassOp

PresentMode

VertexFormat

Traits

AbstractPipeline

Bind
A trait representing a resource that can be bound.

Canvas

Draw

RenderPassExt

RenderTarget
Can be rendered to in a pass.

Renderable

Type Definitions

ScreenTransformation