Module gooey::frontends::renderers::kludgine::kludgine::prelude

Convenience module that exports the public interface of Kludgine

Macros

include_aseprite_sprite

include_font

include_texture

Structs

Batch

Color

DeviceId
Identifier of an input device.

Fill

Font
Font provides TrueType Font rendering

InputEvent
An Event from a device

KeyboardInput
Describes a keyboard input event.

Length
A one-dimensional distance, with value represented by T and unit of measurement Unit.

ModifiersState
Represents the current state of the keyboard modifiers

OpenWindow

Path

PathBuilder

PersistentTileProvider
Provides a simple interface for tile maps that have specific bounds

PreparedSpan

RedrawStatus

RequiresInitialization

Runtime
Runtime is designed to consume the main thread. For cross-platform compatibility, ensure that you call Runtime::run from thee main thread.

Scale
A scaling factor between two different units of measurement.

Scene

Shape

SingleWindowApplication

Sprite

SpriteAnimation

SpriteAnimations

SpriteFrame

SpriteMap

SpriteRotation

SpriteSheet

SpriteSource

SpriteSourceSublocation

Stroke

Surround

Target

Text

Texture

Tile
A Tile represents a sprite at an integer offset on the map

TileMap
TileMap renders tiles retrieved from a TileProvider

Touch
Represents a touch event

Unknown

WindowBuilder

Enums

AnimationMode

DeviceEvent
Represents raw hardware events that are not associated with any particular window.

Dimension

ElementState
Describes the input state of a key.

Event
An input Event

EventStatus

Force
Describes the force of a touch event

KludgineError

MouseButton
Describes a button of a mouse controller.

MouseScrollDelta
Describes a difference in the mouse scroll wheel state.

PathEvent

StartCause
Describes the reason the event loop is resuming.

TouchPhase
Describes touch-screen input state.

VirtualKeyCode
Symbolic name for a keyboard key.

WindowEvent
Describes an event from a Window.

Traits

Application

OpenableWindow

PersistentMap

PointExt

SizeExt

SpriteCollection

TileProvider
TileProvider is how a TileMap retrieves tiles to render

Window
Trait to implement a Window

WindowCreator

Type Definitions

Angle

AxisId
Identifier for a specific analog axis on some device.

ButtonId
Identifier for a specific button on some device.

ControlPoint

Endpoint

KludgineResult
Alias for Result<T,E> where E is KludgineError

PersistentTileMap
PersistentTileMap is an alias for TileMap

Pixels

Point

Points

Raw

Rect

Scaled

ScanCode
Hardware-dependent keyboard scan code.

ScreenScale

Size

Vector