Macro gooey::define_components

macro_rules! define_components {
    ($($widget:ident { $($(#$doc:tt)* $component:ident($type:ty, $name:expr, $($default:tt)*))* })*) => { ... };
    ($type:ty, . $($path:tt)*) => { ... };
    ($type:ty, |$context:ident| $($expr:tt)*) => { ... };
    ($type:ty, @$path:path) => { ... };
    ($type:ty, contrasting!($bg:ident, $($fg:ident),+ $(,)?)) => { ... };
    ($type:ty, $($expr:tt)*) => { ... };
}

Defines a set of style components for Gooey.

These macros implement NamedComponent and ComponentDefinition for each entry defined. The syntax is:

use gooey::define_components;
use gooey::styles::Dimension;
use gooey::styles::components::{SurfaceColor, TextColor};
use gooey::kludgine::Color;

define_components! {
    GroupName {
        /// This is the documentation for example component. It has a default value of `Dimension::ZERO`.
        ExampleComponent(Dimension, "example_component", Dimension::ZERO)
        /// This component whose default value is a color from the current theme.
        ThemedComponent(Color, "themed_component", .primary.color)
        /// This component is a color whose default value is the currently defined `TextColor`.
        DependentComponent(Color, "dependent_component", @TextColor)
        /// This component defaults to picking a contrasting color between `TextColor` and `SurfaceColor`
        ContrastingColor(Color, "contrasting_color", contrasting!(ThemedComponent, TextColor, SurfaceColor))
        /// This component shows how to use a closure for nearly infinite flexibility in computing the default value.
        ClosureDefaultComponent(Color, "closure_component", |context| context.get(&TextColor))
    }
}