Trait gooey::style::Palette

pub trait Palette {
    fn background() -> ColorPair;
    fn foreground() -> ColorPair;
    fn primary() -> ColorPair;
    fn secondary() -> ColorPair;

    fn control_background() -> ColorPair { ... }
    fn navigator_background() -> ColorPair { ... }
    fn navigator_button() -> ColorPair { ... }
}

Returns a set of colors that form a palette.

Required Methods

fn background() -> ColorPair

The background color. Used as the window background color, but also used to derive other neutral background colors.

fn foreground() -> ColorPair

The foreground color. Used as the default text/stroke color, but also used to derive other neutral foreground colors.

fn primary() -> ColorPair

The primary color of the palette. Used to signify default actions, and is considered the primary accent color of the interface.

fn secondary() -> ColorPair

The secondary color of the palette. Used for highlights and information.

Provided Methods

fn control_background() -> ColorPair

The color to use for the background of solid widgets.

fn navigator_background() -> ColorPair

The color for the Navigator bar’s background.

fn navigator_button() -> ColorPair

The color for buttons within the Navigator bar.