Module gooey::styles

Types for styling widgets.

Modules

components

All style components supported by the built-in widgets.

Structs

ColorScheme

A color scheme for a Gooey application.

ColorSchemeBuilder

A builder of ColorSchemes.

ColorSource

A source for [Color]s.

ColorTheme

A pallete of a shared ColorSource.

ComponentName

A fully-qualified style component name.

CustomComponent

A custom component value.

DimensionRange

A range of Dimensions.

Edges

A type describing characteristics about the edges of a rectangle.

FixedTheme

A theme of colors that is shared between light and dark theme variants.

Styles

A collection of style components organized by their name.

SurfaceTheme

A theme of surface colors.

Theme

A Gooey Color theme.

ThemePair

A set of light and dark Themes.

VisualOrder

A 2d ordering configuration.

Enums

Component

A value of a style component.

ContainerLevel

A description of the level of depth a Container is nested at.

Dimension

A 1-dimensional measurement.

FlexibleDimension

A 1-dimensional measurement that may be automatically calculated.

FocusableWidgets

A configuration option to control which controls should be able to receive focus through keyboard focus handling or initial focus handling.

HorizontalOrder

A horizontal direction.

VerticalOrder

A vertical direction.

Traits

ColorExt

Extra functionality added to the [Color] type from Kludgine.

ComponentDefaultvalue

A type that represents a named component with a default value.

ComponentDefinition

A type that represents a named component with a default value of a specific Rust type.

ComponentType

A type that can be converted to and from Component.

IntoComponentValue

A value that can be converted into a Value<Component>.

Lightness

A value that can represent the lightness of a color.

NamedComponent

A type that represents a named style component.

RequireInvalidation

Describes whether a type should invalidate a widget.