Module gooey::styles::components

All style components supported by the built-in widgets.

Structs

AutoFocusableControls

The set of controls to allow focusing via tab key and initial focus selection.

DisabledOutlineColor

A [Color] to be used as an outline color.

Easing

The EasingFunction to apply to animations that have no inherent directionality.

EasingIn

The EasingFunction to apply to animations that transition a value from “nothing” to “something”. For example, if an widget is animating a color’s alpha channel towards opaqueness, it would query for this style component. Otherwise, it would use EasingOut.

EasingOut

The EasingFunction to apply to animations that transition a value from “something” to “nothing”. For example, if an widget is animating a color’s alpha channel towards transparency, it would query for this style component. Otherwise, it would use EasingIn.

HighlightColor

A [Color] to be used as a highlight color.

IntrinsicPadding

Intrinsic, uniform padding for a widget.

LayoutOrder

The VisualOrder strategy to use when laying out content.

LineHeight

The Dimension to use to space multiple lines of text.

OpaqueWidgetColor

A [Color] to be used as a background color for widgets that render an opaque background.

OutlineColor

A [Color] to be used as an outline color.

SurfaceColor

The [Color] of the surface for the user interface to draw upon.

TextColor

The [Color] to use when rendering text.

TextColorVariant

The [Color] to use when rendering text in a more subdued tone.

TextSize

The Dimension to use as the size to render text.

WidgetAccentColor

A [Color] to be used to accent a widget.

WidgetBackground

A [Color] to be used as the background color of a widget.