Trait gooey::styles::ColorExt

pub trait ColorExt: Copy {
    // Required methods
    fn into_source_and_lightness(self) -> (ColorSource, ZeroToOne);
    fn contrast_between(
        self,
        check_source: ColorSource,
        check_lightness: ZeroToOne,
        check_alpha: ZeroToOne
    ) -> ZeroToOne;
    fn most_contrasting(self, others: &[Self]) -> Self
       where Self: Copy;

    // Provided methods
    fn source(self) -> ColorSource { ... }
    fn lightness(self) -> ZeroToOne { ... }
}

Extra functionality added to the [Color] type from Kludgine.

Required Methods

fn into_source_and_lightness(self) -> (ColorSource, ZeroToOne)

Converts this color into its hue/saturation and lightness components.

fn contrast_between( self, check_source: ColorSource, check_lightness: ZeroToOne, check_alpha: ZeroToOne ) -> ZeroToOne

Returns the contrast between this color and the components provided.

To achieve a contrast of 1.0:

self’s hue and check_source.hue must be 180 degrees apart.

self’s saturation and check_source.saturation must be different by 1.0.

self’s lightness and check_lightness must be different by 1.0.

self’s alpha and check_alpha must be different by 1.0.

The algorithm currently used is purposely left undocumented as it will likely change. It should be a reasonable heuristic until someone smarter than @ecton comes along.

fn most_contrasting(self, others: &[Self]) -> Selfwhere Self: Copy,

Returns the color in others that contrasts the most from self.

Provided Methods

fn source(self) -> ColorSource

Returns the hue and saturation of this color.

fn lightness(self) -> ZeroToOne

Returns the perceived lightness of this color.