Trait gooey::styles::ProtoColor

pub trait ProtoColor: Sized {
    // Required methods
    fn hue(&self) -> OklabHue;
    fn saturation(&self) -> Option<ZeroToOne>;

    // Provided method
    fn into_source(
        self,
        saturation_if_not_provided: impl Into<ZeroToOne>
    ) -> ColorSource { ... }
}

A type that can be interpretted as a hue or hue and saturation.

Required Methods

fn hue(&self) -> OklabHue

Returns the hue of this prototype color.

fn saturation(&self) -> Option<ZeroToOne>

Returns the saturation of this prototype color, if available.

Provided Methods

fn into_source( self, saturation_if_not_provided: impl Into<ZeroToOne> ) -> ColorSource

Returns a color source built from this prototype color