Struct gooey::text::rich::RichText

pub struct RichText { /* fields omitted */ }

A multi-paragraph rich text data type.

Implementations

impl RichText

pub fn new(paragraphs: Vec<Text, Global>) -> RichText

Creates a new instance with paragraphs.

pub fn remove_range(&self, range: Range<RichTextPosition>)

Removes a range of characters.

Panics

Panics if range.end is less than range.start.

pub fn insert_str(&self, location: RichTextPosition, value: &str)

Inserts value at location, preserving the style at the location.

pub fn for_each_in_range<F>(&self, range: Range<RichTextPosition>, callback: F) where
F: FnMut(&Text, Range<usize>),

Interates over each paragraph calling callback with the paragraph and the local character range.

pub fn for_each_in_range_mut<F>(
&self,
range: Range<RichTextPosition>,
callback: F
) where
F: Fn(&mut Text, Range<usize>, usize) -> ParagraphRemoval,

Interates over each paragraph calling callback with a mutable reference to the paragraph and the local character range.

pub fn prepare<R>(
&self,
renderer: &R,
wrapping: TextWrap
) -> Vec<PreparedText, Global> where
R: Renderer,

Prepares the rich text for rendering.

pub fn position_after(&self, position: RichTextPosition) -> RichTextPosition

Returns the next RichTextPosition after position. Returns the passed in location if it’s at the end of this text.

pub fn position_before(&self, position: RichTextPosition) -> RichTextPosition

Returns the previous RichTextPosition before position. Returns the first position if it’s at the beginning of this text.

pub fn end(&self) -> RichTextPosition

Returns the last valid location in this text.

pub fn paragraphs(&self) -> Vec<Text, Global>

Returns a clone of the paragraphs contained in this text. To access the data without a copy, use Self::for_each_in_range() or Self::for_each_in_range_mut.

pub fn character_offset_of(&self, position: RichTextPosition) -> usize

Returns the character offset of position within this text. If the position is beyond the bounds of the text, the last character position is returned.

pub fn position_of_character(&self, offset: usize) -> RichTextPosition

Returns the position of the character offset into this text. If the character offset is beyond the bounds of this text, the last position is returned.