Module gooey::value

Types for storing and interacting with values in Widgets.

Structs

BlockUntilUpdatedFuture

Suspends the current async task until the contained value has been updated or there are no remaining writers for the value.

CallbackHandle

A handle to a callback installed on a Dynamic. When dropped, the callback will be uninstalled.

Dynamic

An instance of a value that provides APIs to observe and react to its contents.

DynamicGuard

An exclusive reference to the contents of a Dynamic.

DynamicReader

A reader that tracks the last generation accessed through this reader.

Generation

A tag that represents an individual revision of a Dynamic value.

GenerationalValue

A value stored in a Dynamic with its Generation.

Validations

A grouping of validations that can be checked simultaneously.

WhenValidation

A builder for validations that only run when a precondition is met.

Enums

ReplaceError

An error occurred while updating a value in a Dynamic.

Validation

The status of validating data.

Value

A value that may be either constant or dynamic.

Traits

ForEach

A type that can have a for_each operation applied to it.

ForEachCloned

A type that can have a for_each operation applied to it.

GetWidget

A collection of widgets that can be queried by Key.

IntoDynamic

A type that can convert into a Dynamic<T>.

IntoValue

A type that can be converted into a Value.

MapEach

A type that can create a Dynamic<U> from a T passed into a mapping function.

MapEachCloned

A type that can create a Dynamic<U> from a T passed into a mapping function.

Switchable

A type that can be the source of a Switcher widget.