Enum gooey::value::Value

pub enum Value<T> {
    Constant(T),
    Dynamic(Dynamic<T>),
}

A value that may be either constant or dynamic.

Variants

Constant(T)

A value that will not ever change externally.

Dynamic(Dynamic<T>)

A value that may be updated externally.

Implementations

impl<T> Value<T>

pub fn dynamic(value: T) -> Self

Returns a Value::Dynamic containing value.

pub fn map<R>(&self, map: impl FnOnce(&T) -> R) -> R

Maps the current contents to map and returns the result.

pub fn map_tracking_redraw<R>( &self, context: &WidgetContext<'_, '_>, map: impl FnOnce(&T) -> R ) -> R

Maps the current contents to map and returns the result.

If self is a dynamic, context will be invalidated when the value is updated.

pub fn map_tracking_invalidate<R>( &self, context: &WidgetContext<'_, '_>, map: impl FnOnce(&T) -> R ) -> R

Maps the current contents to map and returns the result.

If self is a dynamic, context will be invalidated when the value is updated.

pub fn map_mut<R>(&mut self, map: impl FnOnce(&mut T) -> R) -> R

Maps the current contents with exclusive access and returns the result.

pub fn map_each<R, F>(&self, map: F) -> Value<R>where F: for<'a> FnMut(&'a T) -> R + Send + 'static, R: Send + 'static,

Returns a new value that is updated using U::from(T.clone()) each time self is updated.

pub fn get(&self) -> Twhere T: Clone,

Returns a clone of the currently stored value.

pub fn get_tracked(&self, context: &WidgetContext<'_, '_>) -> Twhere T: Clone,

Returns a clone of the currently stored value.

If self is a dynamic, context will be invalidated when the value is updated.

pub fn generation(&self) -> Option<Generation>

Returns the current generation of the data stored, if the contained value is Dynamic.

pub fn redraw_when_changed(&self, context: &WidgetContext<'_, '_>)

Marks the widget for redraw when this value is updated.

This function has no effect if the value is constant.

pub fn invalidate_when_changed(&self, context: &WidgetContext<'_, '_>)

Marks the widget for redraw when this value is updated.

This function has no effect if the value is constant.