Trait gooey::value::MapEach

pub trait MapEach<T, U> {
    type Ref<'a>;

    // Required method
    fn map_each<F>(&self, map_each: F) -> Dynamic<U>
       where F: for<'a> FnMut(Self::Ref<'a>) -> U + Send + 'static;
}

A type that can create a Dynamic<U> from a T passed into a mapping function.

Required Associated Types

type Ref<'a>

The borrowed representation of T passed into the mapping function.

Required Methods

fn map_each<F>(&self, map_each: F) -> Dynamic<U>where F: for<'a> FnMut(Self::Ref<'a>) -> U + Send + 'static,

Apply map_each to each value in self, storing the result in the returned dynamic.