Module gooey::widget

Types for creating reusable widgets (aka components or views).

Structs

Callback

A function that can be invoked with a parameter (T) and returns R.

Children

A list of Widgets.

EventHandled

A marker type that represents a handled event.

EventIgnored

A marker type that represents an ignored event.

ManagedWidget

A Widget that has been attached to a widget hierarchy.

WidgetGuard

Exclusive access to a widget.

WidgetId

The unique id of a WidgetInstance.

WidgetInstance

An instance of a Widget.

WidgetTag

A WidgetId that has not been assigned to a WidgetInstance.

WrappedLayout

The layout of a wrapped child widget.

Enums

RootBehavior

A behavior that should be applied to a root widget.

WidgetRef

A child widget

Constants

HANDLED

An EventHandling value that represents a handled event.

IGNORED

An EventHandling value that represents an ignored event.

Traits

MakeWidget

A type that can create a WidgetInstance.

MakeWidgetWithId

A type that can create a WidgetInstance with a preallocated WidgetId.

Widget

A type that makes up a graphical user interface.

WrapperWidget

A Widget that contains a single child.

Type Aliases

EventHandling

A type that represents whether an event has been handled or ignored.