Struct gooey::widget::WidgetInstance

pub struct WidgetInstance { /* private fields */ }

An instance of a Widget.

Implementations

impl WidgetInstance

pub fn new<W>(widget: W) -> Selfwhere W: Widget,

Returns a new instance containing widget.

pub fn id(&self) -> WidgetId

Returns the unique id of this widget instance.

pub fn with_next_focus( self, next_focus: impl IntoValue<Option<WidgetId>> ) -> WidgetInstance

Sets the widget that should be focused next.

Gooey automatically determines reverse tab order by using this same relationship.

pub fn lock(&self) -> WidgetGuard<'_>

Locks the widget for exclusive access. Locking widgets should only be done for brief moments of time when you are certain no deadlocks can occur due to other widget locks being held.

pub fn run(self) -> Result

Runs this widget instance as an application.