Trait gooey::widget::MakeWidget

pub trait MakeWidget: Sized {

    // Required method
    fn make_widget(self) -> WidgetInstance;

    // Provided methods
    fn into_window(self) -> Window<WidgetInstance> { ... }
    fn with_styles(self, styles: impl IntoValue<Styles>) -> Style
       where Self: Sized { ... }
    fn with<C: ComponentDefinition>(
        self,
        name: &C,
        component: impl IntoValue<C::ComponentType>
    ) -> Style
       where Value<C::ComponentType>: IntoComponentValue { ... }
    fn with_next_focus(
        self,
        next_focus: impl IntoValue<Option<WidgetId>>
    ) -> WidgetInstance { ... }
    fn with_enabled(self, enabled: impl IntoValue<bool>) -> WidgetInstance { ... }
    fn into_default(self) -> WidgetInstance { ... }
    fn into_escape(self) -> WidgetInstance { ... }
    fn and(self, other: impl MakeWidget) -> Children { ... }
    fn expand(self) -> Expand { ... }
    fn expand_weighted(self, weight: u8) -> Expand { ... }
    fn expand_horizontally(self) -> Expand { ... }
    fn expand_vertically(self) -> Expand { ... }
    fn size<T>(self, size: Size<T>) -> Resize
       where T: Into<DimensionRange> { ... }
    fn width(self, width: impl Into<DimensionRange>) -> Resize { ... }
    fn height(self, height: impl Into<DimensionRange>) -> Resize { ... }
    fn into_button(self) -> Button { ... }
    fn into_checkbox(self, value: impl IntoDynamic<CheckboxState>) -> Checkbox { ... }
    fn centered(self) -> Align { ... }
    fn align_left(self) -> Align { ... }
    fn align_right(self) -> Align { ... }
    fn align_top(self) -> Align { ... }
    fn align_bottom(self) -> Align { ... }
    fn fit_horizontally(self) -> Align { ... }
    fn fit_vertically(self) -> Align { ... }
    fn scroll(self) -> Scroll { ... }
    fn vertical_scroll(self) -> Scroll { ... }
    fn horizontal_scroll(self) -> Scroll { ... }
    fn widget_ref(self) -> WidgetRef { ... }
    fn contain(self) -> Container { ... }
    fn contain_level(self, level: impl IntoValue<ContainerLevel>) -> Container { ... }
    fn background_color(self, color: impl IntoValue<Color>) -> Container { ... }
    fn pad(self) -> Container { ... }
    fn pad_by(self, padding: impl IntoValue<Edges<Dimension>>) -> Container { ... }
    fn themed(self, theme: impl IntoValue<ThemePair>) -> Themed { ... }
    fn themed_mode(self, mode: impl IntoValue<ThemeMode>) -> ThemedMode { ... }
    fn collapse_horizontally(
        self,
        collapse_when: impl IntoDynamic<bool>
    ) -> Collapse { ... }
    fn collapse_vertically(
        self,
        collapse_when: impl IntoDynamic<bool>
    ) -> Collapse { ... }

}

A type that can create a WidgetInstance.

Required Methods

fn make_widget(self) -> WidgetInstance

Returns a new widget.

Provided Methods

fn into_window(self) -> Window<WidgetInstance>

Returns a new window containing self as the root widget.

fn with_styles(self, styles: impl IntoValue<Styles>) -> Stylewhere Self: Sized,

Associates styles with this widget.

This is equivalent to Style::new(styles, self).

fn with<C: ComponentDefinition>( self, name: &C, component: impl IntoValue<C::ComponentType> ) -> Stylewhere Value<C::ComponentType>: IntoComponentValue,

Associates a style component with self.

fn with_next_focus( self, next_focus: impl IntoValue<Option<WidgetId>> ) -> WidgetInstance

Sets the widget that should be focused next.

Gooey automatically determines reverse tab order by using this same relationship.

fn with_enabled(self, enabled: impl IntoValue<bool>) -> WidgetInstance

Sets this widget to be enabled/disabled based on enabled and returns self.

If this widget is disabled, all children widgets will also be disabled.

Panics

This function can only be called when one instance of the widget exists. If any clones exist, a panic will occur.

fn into_default(self) -> WidgetInstance

Sets this widget as a “default” widget.

Default widgets are automatically activated when the user signals they are ready for the default action to occur.

Example widgets this is used for are:

Submit buttons on forms

Ok buttons

fn into_escape(self) -> WidgetInstance

Sets this widget as an “escape” widget.

Escape widgets are automatically activated when the user signals they are ready to escape their current situation.

Example widgets this is used for are:

Close buttons

Cancel buttons

fn and(self, other: impl MakeWidget) -> Children

Returns a collection of widgets using self and other.

fn expand(self) -> Expand

Expands self to grow to fill its parent.

fn expand_weighted(self, weight: u8) -> Expand

Expands self to grow to fill its parent proportionally with other weighted siblings.

fn expand_horizontally(self) -> Expand

Expands self to grow to fill its parent horizontally.

fn expand_vertically(self) -> Expand

Expands self to grow to fill its parent vertically.

fn size<T>(self, size: Size<T>) -> Resizewhere T: Into<DimensionRange>,

Resizes self to size.

fn width(self, width: impl Into<DimensionRange>) -> Resize

Resizes self to width.

width can be an any of:

Dimension

[Px]

Lp

A range of any fo the above.

fn height(self, height: impl Into<DimensionRange>) -> Resize

Resizes self to height.

height can be an any of:

Dimension

[Px]

Lp

A range of any fo the above.

fn into_button(self) -> Button

Returns this widget as the contents of a clickable button.

fn into_checkbox(self, value: impl IntoDynamic<CheckboxState>) -> Checkbox

Returns this widget as the label of a Checkbox.

fn centered(self) -> Align

Aligns self to the center vertically and horizontally.

fn align_left(self) -> Align

Aligns self to the left.

fn align_right(self) -> Align

Aligns self to the right.

fn align_top(self) -> Align

Aligns self to the top.

fn align_bottom(self) -> Align

Aligns self to the bottom.

fn fit_horizontally(self) -> Align

Fits self horizontally within its parent.

fn fit_vertically(self) -> Align

Fits self vertically within its parent.

fn scroll(self) -> Scroll

Allows scrolling self both vertically and horizontally.

fn vertical_scroll(self) -> Scroll

Allows scrolling self vertically.

fn horizontal_scroll(self) -> Scroll

Allows scrolling self horizontally.

fn widget_ref(self) -> WidgetRef

Creates a WidgetRef for use as child widget.

fn contain(self) -> Container

Wraps self in a Container.

fn contain_level(self, level: impl IntoValue<ContainerLevel>) -> Container

Wraps self in a Container with the specified level.

fn background_color(self, color: impl IntoValue<Color>) -> Container

Returns a new widget that renders color behind self.

fn pad(self) -> Container

Wraps self with the default padding.

fn pad_by(self, padding: impl IntoValue<Edges<Dimension>>) -> Container

Wraps self with the specified padding.

fn themed(self, theme: impl IntoValue<ThemePair>) -> Themed

Applies theme to self and its children.

fn themed_mode(self, mode: impl IntoValue<ThemeMode>) -> ThemedMode

Applies mode to self and its children.

fn collapse_horizontally( self, collapse_when: impl IntoDynamic<bool> ) -> Collapse

Returns a widget that collapses self horizontally based on the dynamic boolean value.

This widget will be collapsed when the dynamic contains true, and revealed when the dynamic contains false.

fn collapse_vertically(self, collapse_when: impl IntoDynamic<bool>) -> Collapse

Returns a widget that collapses self vertically based on the dynamic boolean value.

This widget will be collapsed when the dynamic contains true, and revealed when the dynamic contains false.