Trait gooey::widget::WrapperWidget

pub trait WrapperWidget: Debug + Send + UnwindSafe + 'static {

    // Required method
    fn child_mut(&mut self) -> &mut WidgetRef;

    // Provided methods
    fn layout_child(
        &mut self,
        available_space: Size<ConstraintLimit>,
        context: &mut LayoutContext<'_, '_, '_, '_, '_>
    ) -> Rect<Px> { ... }
    fn mounted(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn unmounted(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn hit_test(
        &mut self,
        location: Point<Px>,
        context: &mut EventContext<'_, '_>
    ) -> bool { ... }
    fn hover(&mut self, location: Point<Px>, context: &mut EventContext<'_, '_>) { ... }
    fn unhover(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn accept_focus(&mut self, context: &mut EventContext<'_, '_>) -> bool { ... }
    fn focus(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn blur(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn activate(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn deactivate(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn mouse_down(
        &mut self,
        location: Point<Px>,
        device_id: DeviceId,
        button: MouseButton,
        context: &mut EventContext<'_, '_>
    ) -> EventHandling { ... }
    fn mouse_drag(
        &mut self,
        location: Point<Px>,
        device_id: DeviceId,
        button: MouseButton,
        context: &mut EventContext<'_, '_>
    ) { ... }
    fn mouse_up(
        &mut self,
        location: Option<Point<Px>>,
        device_id: DeviceId,
        button: MouseButton,
        context: &mut EventContext<'_, '_>
    ) { ... }
    fn keyboard_input(
        &mut self,
        device_id: DeviceId,
        input: KeyEvent,
        is_synthetic: bool,
        context: &mut EventContext<'_, '_>
    ) -> EventHandling { ... }
    fn ime(
        &mut self,
        ime: Ime,
        context: &mut EventContext<'_, '_>
    ) -> EventHandling { ... }
    fn mouse_wheel(
        &mut self,
        device_id: DeviceId,
        delta: MouseScrollDelta,
        phase: TouchPhase,
        context: &mut EventContext<'_, '_>
    ) -> EventHandling { ... }

}

A Widget that contains a single child.

Required Methods

fn child_mut(&mut self) -> &mut WidgetRef

Returns the child widget.

Provided Methods

fn layout_child( &mut self, available_space: Size<ConstraintLimit>, context: &mut LayoutContext<'_, '_, '_, '_, '_> ) -> Rect<Px>

Returns the rectangle that the child widget should occupy given available_space.

fn mounted(&mut self, context: &mut EventContext<'_, '_>)

The widget has been mounted into a parent widget.

fn unmounted(&mut self, context: &mut EventContext<'_, '_>)

The widget has been removed from its parent widget.

fn hit_test( &mut self, location: Point<Px>, context: &mut EventContext<'_, '_> ) -> bool

Returns true if this widget should respond to mouse input at location.

fn hover(&mut self, location: Point<Px>, context: &mut EventContext<'_, '_>)

The widget is currently has a cursor hovering it at location.

fn unhover(&mut self, context: &mut EventContext<'_, '_>)

The widget is no longer being hovered.

fn accept_focus(&mut self, context: &mut EventContext<'_, '_>) -> bool

This widget has been targeted to be focused. If this function returns true, the widget will be focused. If false, Gooey will continue searching for another focus target.

fn focus(&mut self, context: &mut EventContext<'_, '_>)

The widget has received focus for user input.

fn blur(&mut self, context: &mut EventContext<'_, '_>)

The widget is no longer focused for user input.

fn activate(&mut self, context: &mut EventContext<'_, '_>)

The widget has become the active widget.

fn deactivate(&mut self, context: &mut EventContext<'_, '_>)

The widget is no longer active.

fn mouse_down( &mut self, location: Point<Px>, device_id: DeviceId, button: MouseButton, context: &mut EventContext<'_, '_> ) -> EventHandling

A mouse button event has occurred at location. Returns whether the event has been handled or not.

If an event is handled, the widget will receive callbacks for mouse_drag and mouse_up.

fn mouse_drag( &mut self, location: Point<Px>, device_id: DeviceId, button: MouseButton, context: &mut EventContext<'_, '_> )

A mouse button is being held down as the cursor is moved across the widget.

fn mouse_up( &mut self, location: Option<Point<Px>>, device_id: DeviceId, button: MouseButton, context: &mut EventContext<'_, '_> )

A mouse button is no longer being pressed.

fn keyboard_input( &mut self, device_id: DeviceId, input: KeyEvent, is_synthetic: bool, context: &mut EventContext<'_, '_> ) -> EventHandling

A keyboard event has been sent to this widget. Returns whether the event has been handled or not.

fn ime(&mut self, ime: Ime, context: &mut EventContext<'_, '_>) -> EventHandling

An input manager event has been sent to this widget. Returns whether the event has been handled or not.

fn mouse_wheel( &mut self, device_id: DeviceId, delta: MouseScrollDelta, phase: TouchPhase, context: &mut EventContext<'_, '_> ) -> EventHandling

A mouse wheel event has been sent to this widget. Returns whether the event has been handled or not.