Module gooey::widgets

Built-in Widget implementations.

Re-exports

pub use button::Button;

pub use checkbox::Checkbox;

pub use container::Container;

pub use input::Input;

pub use label::Label;

pub use scroll::Scroll;

pub use slider::Slider;

pub use stack::Stack;

Modules

button

A clickable, labeled button

checkbox

A tri-state, labelable checkbox widget.

container

A visual container widget.

input

A text input widget.

label

A read-only text widget.

scroll

A container that scrolls its contents on a virtual surface.

slider

A widget that allows a user to “slide” between values.

stack

A widget that combines a collection of Children widgets into one.

Structs

Align

A widget aligns its contents to its container’s boundaries.

Canvas

A 2d drawable surface.

Expand

A widget that expands its child widget to fill the parent.

Resize

A widget that resizes its contained widget to an explicit size.

Space

A widget that occupies space, optionally filling it with a color.

Style

A widget that applies a set of Styles to all contained widgets.

Switcher

A widget that switches its contents based on a value of T.

Themed

A widget that applies a ThemePair to all contained widgets.

ThemedMode

A widget that applies a set of ThemeMode to all contained widgets.

TileMap

A layered tile-based 2d game surface.