Struct gooey::widgets::grid::GridSection

pub struct GridSection<const N: usize>(/* private fields */);

A single dimension of widgets within a GridWidgets collection.

Implementations

impl GridSection<0>

pub const fn new() -> Self

Returns an empty section.

pub fn and(self, other: impl MakeWidget) -> GridSection<1>

Appends other to the end of this collection of widgets and returns the updated collection.

impl GridSection<1>

pub fn and(self, other: impl MakeWidget) -> GridSection<2>

Appends other to the end of this collection of widgets and returns the updated collection.

impl GridSection<2>

pub fn and(self, other: impl MakeWidget) -> GridSection<3>

Appends other to the end of this collection of widgets and returns the updated collection.

impl GridSection<3>

pub fn and(self, other: impl MakeWidget) -> GridSection<4>

Appends other to the end of this collection of widgets and returns the updated collection.

impl GridSection<4>

pub fn and(self, other: impl MakeWidget) -> GridSection<5>

Appends other to the end of this collection of widgets and returns the updated collection.

impl GridSection<5>

pub fn and(self, other: impl MakeWidget) -> GridSection<6>

Appends other to the end of this collection of widgets and returns the updated collection.

impl GridSection<6>

pub fn and(self, other: impl MakeWidget) -> GridSection<7>

Appends other to the end of this collection of widgets and returns the updated collection.

impl GridSection<7>

pub fn and(self, other: impl MakeWidget) -> GridSection<8>

Appends other to the end of this collection of widgets and returns the updated collection.

impl GridSection<8>

pub fn and(self, other: impl MakeWidget) -> GridSection<9>

Appends other to the end of this collection of widgets and returns the updated collection.

impl GridSection<9>

pub fn and(self, other: impl MakeWidget) -> GridSection<10>

Appends other to the end of this collection of widgets and returns the updated collection.

impl GridSection<10>

pub fn and(self, other: impl MakeWidget) -> GridSection<11>

Appends other to the end of this collection of widgets and returns the updated collection.

impl GridSection<11>

pub fn and(self, other: impl MakeWidget) -> GridSection<12>

Appends other to the end of this collection of widgets and returns the updated collection.