Trait gooey::widgets::input::InputStorage

pub trait InputStorage: UnwindSafe + Send + 'static {
    const MASKED: bool;

    // Required methods
    fn as_str(&self) -> &str;
    fn as_string_mut(&mut self) -> &mut String;
}

A type that can be used as the storage of an Input’s string value.

This crate implements this trait for these types:

String

Cow<'static, str>

CowString

MaskedString

Required Associated Constants

const MASKED: bool

If true, the input field should display a mask instead of the actual string by default.

Required Methods

fn as_str(&self) -> &str

Returns a reference to the contents as a str.

fn as_string_mut(&mut self) -> &mut String

Returns an exclusive reference to the contents as a String.