Trait gooey::widgets::progress::ProgressValue

pub trait ProgressValue: 'static {
    type Value;

    // Required method
    fn to_progress(
        &self,
        range: Option<RangeInclusive<&Self::Value>>
    ) -> Progress;
}

A value that can be used in a progress indicator.

Required Associated Types

type Value

The type that progress is ranged over.

Required Methods

fn to_progress(&self, range: Option<RangeInclusive<&Self::Value>>) -> Progress

Converts this value to a progress using the range given, if provided. If no range is provided, the full range of the type should be considered.