Struct gooey::widgets::select::Select

pub struct Select<T> {
    pub value: T,
    pub state: Dynamic<T>,
    pub kind: Value<ButtonKind>,
    /* private fields */
}

A selectable, labeled widget representing a value.

Fields
value: T

The value this button represents.

state: Dynamic<T>

The state (value) of the select.

kind: Value<ButtonKind>

The button kind to use as the basis for this select. Selects default to ButtonKind::Transparent.

Implementations

impl<T> Select<T>

pub fn new(value: T, state: impl IntoDynamic<T>, label: impl MakeWidget) -> Self

Returns a new select that sets state to value when pressed. label is drawn inside of the button.

pub fn kind(self, kind: impl IntoValue<ButtonKind>) -> Self

Updates the button kind to use as the basis for this select, and returns self.

Selects default to ButtonKind::Transparent.