Struct gooey::widgets::Canvas

pub struct Canvas { /* private fields */ }

A 2d drawable surface.

Implementations

impl Canvas

pub fn new<F>(render: F) -> Selfwhere F: for<'clip, 'gfx, 'pass, 'context, 'window> FnMut(&mut GraphicsContext<'context, 'window, 'clip, 'gfx, 'pass>) + Send + UnwindSafe + 'static,

Returns a new canvas that draws its contents by invoking render.

pub fn tick(self, tick: Tick) -> Self

Associates a Tick with this widget and returns self.