Struct gooey::widgets::Expand

pub struct Expand { /* private fields */ }

A widget that expands its child widget to fill the parent.

Some parent widgets support weighting children when there is more than one Expanded widget.

Implementations

impl Expand

pub fn new(child: impl MakeWidget) -> Self

Returns a widget that expands child to fill the parent widget.

pub fn horizontal(child: impl MakeWidget) -> Self

Returns a widget that expands child to fill the parent widget horizontally.

pub fn vertical(child: impl MakeWidget) -> Self

Returns a widget that expands child to fill the parent widget vertically.

pub fn empty() -> Self

Returns a widget that expands to fill its parent, but has no contents.

pub fn weighted(weight: u8, child: impl MakeWidget) -> Self

Returns a widget that expands child to fill the parent widget, using weight when competing with available space with other Expands.

Note: Not all container widgets support weighted expansion.

pub const fn child(&self) -> &WidgetRef

Returns a reference to the child widget.