Struct gooey::widgets::Input

pub struct Input {
    pub text: Value<String>,
    /* private fields */
}

A text input widget.

Fields
text: Value<String>

The value of this widget.

Implementations

impl Input

pub fn empty() -> Self

Returns an empty widget.

pub fn new(initial_text: impl IntoValue<String>) -> Self

Returns a new widget containing initial_text.

pub fn on_key<F>(self, on_key: F) -> Selfwhere F: FnMut(KeyEvent) -> EventHandling + Send + UnwindSafe + 'static,

Sets the on_key callback.

This function is called for every keyboard input event. If HANDLED is returned, this widget will ignore the event.