Struct gooey::window::Window

pub struct Window<Behavior>where
    Behavior: WindowBehavior,{
    pub attributes: WindowAttributes,
    pub theme: Value<ThemePair>,
    /* private fields */
}

A Gooey window that is not yet running.

Fields
attributes: WindowAttributes

The attributes of this window.

theme: Value<ThemePair>

The colors to use to theme the user interface.

Implementations

impl Window<WidgetInstance>

pub fn for_widget<W>(widget: W) -> Selfwhere W: Widget,

Returns a new instance using widget as its contents.

pub fn with_focused(self, focused: impl IntoDynamic<bool>) -> Self

Sets focused to be the dynamic updated when this window’s focus status is changed.

When the window is focused for user input, the dynamic will contain true.

focused will be initialized with an initial state of false.

pub fn with_occluded(self, occluded: impl IntoDynamic<bool>) -> Self

Sets occluded to be the dynamic updated when this window’s occlusion status is changed.

When the window is occluded (completely hidden/offscreen/minimized), the dynamic will contain true. If the window is at least partially visible, this value will contain true.

occluded will be initialized with an initial state of false.

impl<Behavior> Window<Behavior>where Behavior: WindowBehavior,

pub fn new(context: Behavior::Context) -> Self

Returns a new instance using context to initialize the window upon opening.