Trait gooey::animation::AnimationTarget

pub trait AnimationTarget: Sized + Send + Sync {
    type Running: AnimateTarget;

    // Required method
    fn begin(self) -> Self::Running;

    // Provided method
    fn over(self, duration: Duration) -> Animation<Self, Linear> { ... }
}

A target for a timed Animation.

Required Associated Types

type Running: AnimateTarget

The type that can linearly interpolate this target.

Required Methods

fn begin(self) -> Self::Running

Record the current value of the target, and return a type that can linearly interpolate between the current value and the desired value.

Provided Methods

fn over(self, duration: Duration) -> Animation<Self, Linear>

Returns a pending animation that linearly transitions self over duration.

A different Easing can be used by calling with_easing on the result of this function.