Module gooey::core

Core traits and types used to create Graphical User Interfaces (GUIs - gooeys).

Modules

euclid
A collection of strongly typed math tools for computer graphics with an inclination towards 2d graphics and layout.

palette
A library that makes linear color calculations and conversion easy and accessible for anyone. It uses the type system to enforce correctness and to avoid mistakes, such as mixing incompatible color types.

renderer
Types used for drawing.

styles
Types used for styling.

Structs

AnyTransmogrifierContext
A context used internally when the Transmogrifier type cannot be known.

Callback
A callback that receives information I, and returns R.

Channels
Communication channels used to communicate between Widgets and Transmogrifiers.

Context
Enables Widgets to send commands to the Transmogrifier.

Gooey
A graphical user interface.

Pixels
A unit representing physical pixels on a display.

Points
A unit representing Desktop publishing points/PostScript points. Measurements in this scale are equal to 1/72 of an imperial inch.

StyledWidget
A widget and its initial style information.

TransmogrifierContext
A context passed into Transmogrifier functions with access to useful data and types. This type is mostly used to avoid passing so many parameters across all functions.

TransmogrifierState
Generic storage for a transmogrifier.

Transmogrifiers
A collection of transmogrifiers to use inside of a frontend.

WeakWidgetRegistration
References an initialized widget. These references will not keep a widget from being removed.

WidgetId
A unique ID of a widget, with information about the widget type.

WidgetRef
A widget reference. Does not prevent a widget from being destroyed if removed from an interface.

WidgetRegistration
References an initialized widget. On drop, frees the storage and id.

WidgetState
Generic, clone-able storage for a widget’s transmogrifier.

WidgetStorage
Generic-type-less widget storage.

Constants

ROOT_CLASS
The name of the class assigned to the root widget of a window.

Traits

AnyChannels
A generic-type-less trait for Channels

AnyFrontend
An interface for Frontend that doesn’t requier knowledge of associated types.

AnySendSync
A value that can be used as Any that is threadsafe.

AnyTransmogrifier
A Transmogrifier without any associated types.

AnyWidget
A Widget without any associated types. Useful for implementing frontends.

CallbackFn
A callback implementation. Not typically directly implemented, as this trait is auto-implemented for any Fn(I) -> R types.

Frontend
A frontend is an implementation of widgets and layouts.

Transmogrifier
Transforms a Widget into whatever is needed for Frontend F.

Widget
A graphical user interface element.