Struct gooey::core::palette::Alpha

#[repr(C)]pub struct Alpha<C, T> {
    pub color: C,
    pub alpha: T,
}

An alpha component wrapper for colors.

Fields
color: C

The color.

alpha: T

The transparency component. 0.0 is fully transparent and 1.0 is fully opaque.

Implementations

impl<T, A> Alpha<Hsl<Srgb, T>, A> where
T: Component + Float,
A: Component,

Hsla implementations.

pub fn new<H>(
hue: H,
saturation: T,
lightness: T,
alpha: A
) -> Alpha<Hsl<Srgb, T>, A> where
H: Into<RgbHue<T>>,

HSL and transparency for linear sRGB.

impl<S, T, A> Alpha<Hsl<S, T>, A> where
T: Component + Float,
S: RgbSpace,
A: Component,

Hsla implementations.

pub fn with_wp<H>(
hue: H,
saturation: T,
lightness: T,
alpha: A
) -> Alpha<Hsl<S, T>, A> where
H: Into<RgbHue<T>>,

Linear HSL and transparency.

pub fn into_components(self) -> (RgbHue<T>, T, T, A)

Convert to a (hue, saturation, lightness, alpha) tuple.

pub fn from_components<H>((H, T, T, A)) -> Alpha<Hsl<S, T>, A> where
H: Into<RgbHue<T>>,

Convert from a (hue, saturation, lightness, alpha) tuple.

impl<T, A> Alpha<Hsv<Srgb, T>, A> where
T: Component + Float,
A: Component,

Hsva implementations.

pub fn new<H>(
hue: H,
saturation: T,
value: T,
alpha: A
) -> Alpha<Hsv<Srgb, T>, A> where
H: Into<RgbHue<T>>,

HSV and transparency for linear sRGB.

impl<S, T, A> Alpha<Hsv<S, T>, A> where
T: Component + Float,
S: RgbSpace,
A: Component,

Hsva implementations.

pub fn with_wp<H>(
hue: H,
saturation: T,
value: T,
alpha: A
) -> Alpha<Hsv<S, T>, A> where
H: Into<RgbHue<T>>,

Linear HSV and transparency.

pub fn into_components(self) -> (RgbHue<T>, T, T, A)

Convert to a (hue, saturation, value, alpha) tuple.

pub fn from_components<H>((H, T, T, A)) -> Alpha<Hsv<S, T>, A> where
H: Into<RgbHue<T>>,

Convert from a (hue, saturation, value, alpha) tuple.

impl<T, A> Alpha<Hwb<Srgb, T>, A> where
T: Component + Float,
A: Component,

Hwba implementations.

pub fn new<H>(
hue: H,
whiteness: T,
blackness: T,
alpha: A
) -> Alpha<Hwb<Srgb, T>, A> where
H: Into<RgbHue<T>>,

HWB and transparency for linear sRGB.

impl<S, T, A> Alpha<Hwb<S, T>, A> where
T: Component + Float,
S: RgbSpace,
A: Component,

Hwba implementations.

pub fn with_wp<H>(
hue: H,
whiteness: T,
blackness: T,
alpha: A
) -> Alpha<Hwb<S, T>, A> where
H: Into<RgbHue<T>>,

Linear HWB and transparency.

pub fn into_components(self) -> (RgbHue<T>, T, T, A)

Convert to a (hue, whiteness, blackness, alpha) tuple.

pub fn from_components<H>((H, T, T, A)) -> Alpha<Hwb<S, T>, A> where
H: Into<RgbHue<T>>,

Convert from a (hue, whiteness, blackness, alpha) tuple.

impl<T, A> Alpha<Lab<D65, T>, A> where
T: Component + Float,
A: Component,

Laba implementations.

pub fn new(l: T, a: T, b: T, alpha: A) -> Alpha<Lab<D65, T>, A>

CIE L*a*b* and transparency and white point D65.

impl<Wp, T, A> Alpha<Lab<Wp, T>, A> where
T: Component + Float,
A: Component,
Wp: WhitePoint,

Laba implementations.

pub fn with_wp(l: T, a: T, b: T, alpha: A) -> Alpha<Lab<Wp, T>, A>

CIE L*a*b* and transparency.

pub fn into_components(self) -> (T, T, T, A)

Convert to a (L\*, a\*, b\*, alpha) tuple.

pub fn from_components((T, T, T, A)) -> Alpha<Lab<Wp, T>, A>

Convert from a (L\*, a\*, b\*, alpha) tuple.

impl<T, A> Alpha<Lch<D65, T>, A> where
T: Component + Float,
A: Component,

Lcha implementations.

pub fn new<H>(l: T, chroma: T, hue: H, alpha: A) -> Alpha<Lch<D65, T>, A> where
H: Into<LabHue<T>>,

CIE L*C*h° and transparency with white point D65.

impl<Wp, T, A> Alpha<Lch<Wp, T>, A> where
T: Component + Float,
A: Component,
Wp: WhitePoint,

Lcha implementations.

pub fn with_wp<H>(l: T, chroma: T, hue: H, alpha: A) -> Alpha<Lch<Wp, T>, A> where
H: Into<LabHue<T>>,

CIE L*C*h° and transparency.

pub fn into_components(self) -> (T, T, LabHue<T>, A)

Convert to a (L\*, C\*, h°, alpha) tuple.

pub fn from_components<H>((T, T, H, A)) -> Alpha<Lch<Wp, T>, A> where
H: Into<LabHue<T>>,

Convert from a (L\*, C\*, h°, alpha) tuple.

impl<S, T, A> Alpha<Luma<S, T>, A> where
T: Component,
S: LumaStandard,
A: Component,

Lumaa implementations.

pub fn new(luma: T, alpha: A) -> Alpha<Luma<S, T>, A>

Create a luminance color with transparency.

pub fn into_format<U, B>(self) -> Alpha<Luma<S, U>, B> where
U: Component,
B: Component,

Convert into another component type.

pub fn from_format<U, B>(color: Alpha<Luma<S, U>, B>) -> Alpha<Luma<S, T>, A> where
U: Component,
B: Component,

Convert from another component type.

pub fn into_components(self) -> (T, A)

Convert to a (luma, alpha) tuple.

pub fn from_components((T, A)) -> Alpha<Luma<S, T>, A>

Convert from a (luma, alpha) tuple.

impl<S, T, A> Alpha<Luma<S, T>, A> where
T: Component + Float,
S: LumaStandard,
A: Component,

Lumaa implementations.

pub fn into_linear(
self
) -> Alpha<Luma<Linear<<S as LumaStandard>::WhitePoint>, T>, A>

Convert the color to linear luminance with transparency.

pub fn from_linear(
color: Alpha<Luma<Linear<<S as LumaStandard>::WhitePoint>, T>, A>
) -> Alpha<Luma<S, T>, A>

Convert linear luminance to nonlinear luminance with transparency.

pub fn into_encoding<St>(self) -> Alpha<Luma<St, T>, A> where
St: LumaStandard<WhitePoint = <S as LumaStandard>::WhitePoint>,

Convert the color to a different encoding with transparency.

pub fn from_encoding<St>(color: Alpha<Luma<St, T>, A>) -> Alpha<Luma<S, T>, A> where
St: LumaStandard<WhitePoint = <S as LumaStandard>::WhitePoint>,

Convert luminance from a different encoding with transparency.

impl<S, T, A> Alpha<Rgb<S, T>, A> where
T: Component,
S: RgbStandard,
A: Component,

Rgba implementations.

pub fn new(red: T, green: T, blue: T, alpha: A) -> Alpha<Rgb<S, T>, A>

Nonlinear RGB.

pub fn into_format<U, B>(self) -> Alpha<Rgb<S, U>, B> where
U: Component,
B: Component,

Convert into another component type.

pub fn from_format<U, B>(color: Alpha<Rgb<S, U>, B>) -> Alpha<Rgb<S, T>, A> where
U: Component,
B: Component,

Convert from another component type.

pub fn into_components(self) -> (T, T, T, A)

Convert to a (red, green, blue, alpha) tuple.

pub fn from_components((T, T, T, A)) -> Alpha<Rgb<S, T>, A>

Convert from a (red, green, blue, alpha) tuple.

impl<S, T, A> Alpha<Rgb<S, T>, A> where
T: Component + Float,
S: RgbStandard,
A: Component,

Rgba implementations.

pub fn into_linear(self) -> Alpha<Rgb<Linear<<S as RgbStandard>::Space>, T>, A>

Convert the color to linear RGB with transparency.

pub fn from_linear(
color: Alpha<Rgb<Linear<<S as RgbStandard>::Space>, T>, A>
) -> Alpha<Rgb<S, T>, A>

Convert linear RGB to nonlinear RGB with transparency.

pub fn into_encoding<St>(self) -> Alpha<Rgb<St, T>, A> where
St: RgbStandard<Space = <S as RgbStandard>::Space>,

Convert the color to a different encoding with transparency.

pub fn from_encoding<St>(color: Alpha<Rgb<St, T>, A>) -> Alpha<Rgb<S, T>, A> where
St: RgbStandard<Space = <S as RgbStandard>::Space>,

Convert RGB from a different encoding with transparency.

impl<T, A> Alpha<Xyz<D65, T>, A> where
T: Component + Float,
A: Component,

Xyza implementations.

pub fn new(x: T, y: T, luma: T, alpha: A) -> Alpha<Xyz<D65, T>, A>

CIE Yxy and transparency with white point D65.

impl<Wp, T, A> Alpha<Xyz<Wp, T>, A> where
T: Component + Float,
A: Component,
Wp: WhitePoint,

Xyza implementations.

pub fn with_wp(x: T, y: T, z: T, alpha: A) -> Alpha<Xyz<Wp, T>, A>

CIE XYZ and transparency.

pub fn into_components(self) -> (T, T, T, A)

Convert to a (X, Y, Z, alpha) tuple.

pub fn from_components((T, T, T, A)) -> Alpha<Xyz<Wp, T>, A>

Convert from a (X, Y, Z, alpha) tuple.

impl<T, A> Alpha<Yxy<D65, T>, A> where
T: Component + Float,
A: Component,

Yxya implementations.

pub fn new(x: T, y: T, luma: T, alpha: A) -> Alpha<Yxy<D65, T>, A>

CIE Yxy and transparency with white point D65.

impl<Wp, T, A> Alpha<Yxy<Wp, T>, A> where
T: Component + Float,
A: Component,
Wp: WhitePoint,

Yxya implementations.

pub fn with_wp(x: T, y: T, luma: T, alpha: A) -> Alpha<Yxy<Wp, T>, A>

CIE Yxy and transparency.

pub fn into_components(self) -> (T, T, T, A)

Convert to a (x, y, luma), a.k.a. (x, y, Y) tuple.

pub fn from_components((T, T, T, A)) -> Alpha<Yxy<Wp, T>, A>

Convert from a (x, y, luma), a.k.a. (x, y, Y) tuple.