Trait gooey::core::Frontend

pub trait Frontend: 'static + Clone + Debug + Send + Sync {
    type AnyTransmogrifier: AnyTransmogrifier<Self>;
    type Context;
    fn gooey(&self) -> &Gooey<Self>;

    fn theme(&self) -> SystemTheme;

    fn set_widget_has_messages(&self, widget: WidgetId);

    fn ui_state_for(&self, widget_id: &WidgetId) -> State { ... }

    fn exit_managed_code(&self) { ... }
}

A frontend is an implementation of widgets and layouts.

Associated Types

type AnyTransmogrifier: AnyTransmogrifier<Self>

The generic-free type of the frontend-specific transmogrifier trait.

type Context

The context type provided to aide in transmogrifying.

Required methods

fn gooey(&self) -> &Gooey<Self>

Returns the underlying Gooey instance.

fn theme(&self) -> SystemTheme

Returns the current system theme.

fn set_widget_has_messages(&self, widget: WidgetId)

Notifies the frontend that a widget has messages. Frontends should ensure that process_widget_messages is called at some point after this method is called.

Provided methods

fn ui_state_for(&self, widget_id: &WidgetId) -> State

Returns the State for the widget. Not all frontends track UI state, and this function is primarily used when building frontends.

fn exit_managed_code(&self)

Executed when Gooey exits a managed code block.