Trait gooey::core::RelatedStorage

pub trait RelatedStorage<K>: 'static + Debug + Send + Sync where
    K: Key, {
    fn widget(&self) -> WeakWidgetRegistration;

    fn remove(&self, key: &K) -> Option<WeakWidgetRegistration>;

    fn register(&self, key: K, widget: &WidgetRegistration);
}

Related storage enables a widget to communicate in a limited way about widgets being inserted or removed.

Required methods

fn widget(&self) -> WeakWidgetRegistration

Returns the registration of the widget that this is from.

fn remove(&self, key: &K) -> Option<WeakWidgetRegistration>

Removes the widget with key from this storage. Returns the removed registration if one was removed.

fn register(&self, key: K, widget: &WidgetRegistration)

Registers widget with key.