Trait gooey::core::Transmogrifier

pub trait Transmogrifier<F>: Send + Sync where
    F: Frontend, {
    type Widget: Widget;
    type Context: Send + Sync;
    pub fn content_size(
        &self, 
        widget: &Self::Widget, 
        constraints: Size2D<Option<f32>, Points>, 
        context: &Self::Context
    ) -> Size2D<f32, Points>;
}

Transforms a Widget into whatever is needed for Frontend F.

Associated Types

type Widget: Widget

The type of the widget being transmogrified.

type Context: Send + Sync

The frontend-specific context type provided to aide in transmogrifying.

Loading content...
Required methods

pub fn content_size(
&self,
widget: &Self::Widget,
constraints: Size2D<Option<f32>, Points>,
context: &Self::Context
) -> Size2D<f32, Points>

Calculate the content-size needed for this widget, trying to stay within constraints.

Loading content...