Trait gooey::core::Transmogrifier

pub trait Transmogrifier<F>: Debug where
    F: Frontend, {
    type Widget: Widget;
    type State: Default + Debug + Any + Send + Sync;
    pub fn initialize(
        &self, 
        widget: &Self::Widget, 
        reference: &WidgetRef<Self::Widget>, 
        frontend: &F
    ) -> Self::State { ... }

    pub fn receive_command(
        &self, 
        state: &mut Self::State, 
        command: <Self::Widget as Widget>::TransmogrifierCommand, 
        widget: &Self::Widget, 
        frontend: &F
    ) { ... }

    pub fn process_messages(
        &self, 
        state: &mut Self::State, 
        widget: &mut Self::Widget, 
        frontend: &F, 
        channels: &Channels<Self::Widget>
    ) { ... }

    pub fn default_state_for(
        &self, 
        widget: &Self::Widget, 
        reference: &WidgetRef<Self::Widget>, 
        frontend: &F
    ) -> TransmogrifierState { ... }

    pub fn widget_type_id(&self) -> TypeId { ... }
}

Transforms a Widget into whatever is needed for Frontend F.

Associated Types

type Widget: Widget

The type of the widget being transmogrified.

type State: Default + Debug + Any + Send + Sync

The type the storage this transmogrifier uses for state.

Loading content...
Provided methods

pub fn initialize(
&self,
widget: &Self::Widget,
reference: &WidgetRef<Self::Widget>,
frontend: &F
) -> Self::State

Called after a transmogrifier is initialized.

pub fn receive_command(
&self,
state: &mut Self::State,
command: <Self::Widget as Widget>::TransmogrifierCommand,
widget: &Self::Widget,
frontend: &F
)

Called when a command is received from the widget.

pub fn process_messages(
&self,
state: &mut Self::State,
widget: &mut Self::Widget,
frontend: &F,
channels: &Channels<Self::Widget>
)

Processes commands and events for this widget and transmogrifier.

pub fn default_state_for(
&self,
widget: &Self::Widget,
reference: &WidgetRef<Self::Widget>,
frontend: &F
) -> TransmogrifierState

Returns an initialized state using Self::State::default().

pub fn widget_type_id(&self) -> TypeId

Returns the TypeId of Self::Widget.

Loading content...