Trait gooey::core::Widget

pub trait Widget: 'static + Debug + Send + Sync {
    type Command: Debug + Send + Sync;
    type Event: Debug + Send + Sync;

    pub const CLASS: &'static str;
    #[must_use]
    pub fn classes() -> Classes { ... }

    pub fn receive_event(&mut self, event: Self::Event, context: &Context<Self>) { ... }

    #[must_use]
    pub fn text_color(style: &Style) -> Option<&ColorPair> { ... }

    #[must_use]
    pub fn background_color(style: &Style) -> Option<&ColorPair> { ... }

    pub fn map<OW, F, R>(
        widget_id: &WidgetId, 
        context: &Context<Self>, 
        with_fn: F
    ) -> Option<R>
    where
        F: FnOnce(&OW, &Context<OW>) -> R,
        OW: Widget,
    { ... }

    pub fn map_mut<OW, F, R>(
        widget_id: &WidgetId, 
        context: &Context<Self>, 
        with_fn: F
    ) -> Option<R>
    where
        F: FnOnce(&mut OW, &Context<OW>) -> R,
        OW: Widget,
    { ... }
}

A graphical user interface element.

Associated Types

type Command: Debug + Send + Sync

Widgets may need to communicate with transmogrifier implementations. This type is the type that can be sent to a transmogrifier.

type Event: Debug + Send + Sync

The type of the event that any Transmogrifier for this widget to use.

Loading content...
Associated Constants

pub const CLASS: &'static str

The unique class name for this widget. Must not conflict with any other widgets in use. Widget authors should prefix their widget names to ensure no conflicts. For example, the gooey-widgets crate prefixes all of the CLASS constants with gooey-.

Loading content...
Provided methods

#[must_use]
pub fn classes() -> Classes

Returns all classes that apply styles for this widget.

pub fn receive_event(&mut self, event: Self::Event, context: &Context<Self>)

Called when an event from the transmogrifier was received.

#[must_use]
pub fn text_color(style: &Style) -> Option<&ColorPair>

Returns the effective text color for the given style.

#[must_use]
pub fn background_color(style: &Style) -> Option<&ColorPair>

Returns the effective background color for the given style.

pub fn map<OW, F, R>(
widget_id: &WidgetId,
context: &Context<Self>,
with_fn: F
) -> Option<R> where
F: FnOnce(&OW, &Context<OW>) -> R,
OW: Widget,

Invokes with_fn with the widget widget_id and a Context. Returns the result.

Returns None if OW does not match the type of the widget contained.

pub fn map_mut<OW, F, R>(
widget_id: &WidgetId,
context: &Context<Self>,
with_fn: F
) -> Option<R> where
F: FnOnce(&mut OW, &Context<OW>) -> R,
OW: Widget,

Invokes with_fn with the widget widget_id and a Context. Returns the result.

Returns None if OW does not match the type of the widget contained.

Loading content...