Trait gooey::core::Widget

pub trait Widget: 'static + Debug + Send + Sync {
    type Command: Debug + Send + Sync;
    type TransmogrifierCommand: Debug + Send + Sync;
    type TransmogrifierEvent: Debug + Send + Sync;

    pub const CLASS: &'static str;
    pub fn receive_event(
        &mut self, 
        event: Self::TransmogrifierEvent, 
        context: &Context<Self>
    ) { ... }

    pub fn receive_command(
        &mut self, 
        command: Self::Command, 
        context: &Context<Self>
    ) { ... }
}

A graphical user interface element.

Associated Types

type Command: Debug + Send + Sync

Widgets may need to receive instructions from other entities. This type is the type other widgets can use to communicate with this widget;

type TransmogrifierCommand: Debug + Send + Sync

Widgets may need to communicate with transmogrifier implementations. This type is the type that can be sent to a transmogrifier.

type TransmogrifierEvent: Debug + Send + Sync

The type of the event that any Transmogrifier for this widget to use.

Loading content...
Associated Constants

pub const CLASS: &'static str

The unique class name for this widget. Must not conflict with any other widgets in use. Widget authors should prefix their widget names to ensure no conflicts. For example, the gooey-widgets crate prefixes all of the CLASS constants with gooey-.

Loading content...
Provided methods

pub fn receive_event(
&mut self,
event: Self::TransmogrifierEvent,
context: &Context<Self>
)

Called when an event from the transmogrifier was received.

pub fn receive_command(
&mut self,
command: Self::Command,
context: &Context<Self>
)

Called when an event from the transmogrifier was received.

Loading content...