Trait gooey::core::Widget

pub trait Widget: 'static + Debug + Send + Sync {
    type Command: Debug + Send + Sync;
    type Event: Debug + Send + Sync;

    const CLASS: &'static str;
    const FOCUSABLE: bool;
    fn classes() -> Classes { ... }

    fn receive_event(&mut self, event: Self::Event, context: &Context<Self>) { ... }

    fn text_color(style: &Style) -> Option<&ColorPair> { ... }

    fn background_color(style: &Style) -> Option<&ColorPair> { ... }

    fn map<OW, F, R>(
        widget_id: &WidgetId, 
        context: &Context<Self>, 
        with_fn: F
    ) -> Option<R>
    where
        F: FnOnce(&OW, &Context<OW>) -> R,
        OW: Widget,
    { ... }

    fn map_mut<OW, F, R>(
        widget_id: &WidgetId, 
        context: &Context<Self>, 
        with_fn: F
    ) -> Option<R>
    where
        F: FnOnce(&mut OW, &Context<OW>) -> R,
        OW: Widget,
    { ... }
}

A graphical user interface element.

Associated Types

type Command: Debug + Send + Sync

Widgets may need to communicate with transmogrifier implementations. This type is the type that can be sent to a transmogrifier.

type Event: Debug + Send + Sync

The type of the event that any Transmogrifier for this widget to use.

Associated Constants

const CLASS: &'static str

The unique class name for this widget. Must not conflict with any other widgets in use. Widget authors should prefix their widget names to ensure no conflicts. For example, the gooey-widgets crate prefixes all of the CLASS constants with gooey-.

const FOCUSABLE: bool

When true, the control is able to receive focus through focus advancement, most commonly done when using the tab key.

Provided methods

fn classes() -> Classes

Returns all classes that apply styles for this widget.

fn receive_event(&mut self, event: Self::Event, context: &Context<Self>)

Called when an event from the transmogrifier was received.

fn text_color(style: &Style) -> Option<&ColorPair>

Returns the effective text color for the given style.

fn background_color(style: &Style) -> Option<&ColorPair>

Returns the effective background color for the given style.

fn map<OW, F, R>(
widget_id: &WidgetId,
context: &Context<Self>,
with_fn: F
) -> Option<R> where
F: FnOnce(&OW, &Context<OW>) -> R,
OW: Widget,

Invokes with_fn with the widget widget_id and a Context. Returns the result.

Returns None if OW does not match the type of the widget contained.

fn map_mut<OW, F, R>(
widget_id: &WidgetId,
context: &Context<Self>,
with_fn: F
) -> Option<R> where
F: FnOnce(&mut OW, &Context<OW>) -> R,
OW: Widget,

Invokes with_fn with the widget widget_id and a Context. Returns the result.

Returns None if OW does not match the type of the widget contained.