Module gooey::styles::components

All style components supported by the built-in widgets.

Structs

AutoFocusableControls

The set of controls to allow focusing via tab key and initial focus selection.

BaseLineHeight

The base Dimension to use to space multiple lines of text. Unless overridden, all other sizes for built-in widgets will be based on this dimension.

BaseTextSize

The base Dimension to use as the normal text size. Unless overridden, all other sizes for built-in widgets will be based on this dimension.

CornerRadius

A set of radius descriptions for how much roundness to apply to the shapes of widgets.

DisabledOutlineColor

A [Color] to be used as an outline color.

DisabledWidgetAccentColor

A [Color] to be used to accent a disabled widget.

Easing

The EasingFunction to apply to animations that have no inherent directionality.

EasingIn

The EasingFunction to apply to animations that transition a value from “nothing” to “something”. For example, if an widget is animating a color’s alpha channel towards opaqueness, it would query for this style component. Otherwise, it would use EasingOut.

EasingOut

The EasingFunction to apply to animations that transition a value from “something” to “nothing”. For example, if an widget is animating a color’s alpha channel towards transparency, it would query for this style component. Otherwise, it would use EasingIn.

FontFamily

The font family to render text using.

FontStyle

The font style to apply to text rendering.

FontWeight

The font (boldness) weight to apply to text rendering.

Heading1FontFamily

The FontFamilyList to apply to h1 headings.

Heading1Style

The [Style] to apply to h1 headings.

Heading1Weight

The [Weight] to apply to h1 headings.

Heading2FontFamily

The FontFamilyList to apply to h2 headings.

Heading2Style

The [Style] to apply to h2 headings.

Heading2Weight

The [Weight] to apply to h2 headings.

Heading3FontFamily

The FontFamilyList to apply to h3 headings.

Heading3Style

The [Style] to apply to h3 headings.

Heading3Weight

The [Weight] to apply to h3 headings.

Heading4FontFamily

The FontFamilyList to apply to h4 headings.

Heading4Style

The [Style] to apply to h4 headings.

Heading4Weight

The [Weight] to apply to h4 headings.

Heading5FontFamily

The FontFamilyList to apply to h5 headings.

Heading5Style

The [Style] to apply to h5 headings.

Heading5Weight

The [Weight] to apply to h5 headings.

Heading6FontFamily

The FontFamilyList to apply to h6 headings.

Heading6Style

The [Style] to apply to h6 headings.

Heading6Weight

The [Weight] to apply to h6 headings.

HeadingFontFamily

The default FontFamilyList to apply to headings.

HeadingStyle

The default [Style] to apply to headings.

HeadingWeight

The default [Weight] to apply to headings.

HighlightColor

A [Color] to be used as a highlight color.

IntrinsicPadding

Intrinsic, uniform padding for a widget.

LayoutOrder

The VisualOrder strategy to use when laying out content.

LineHeight

The Dimension to use to space multiple lines of text.

LineHeight1

The smallest line height on a series of 8 steps.

LineHeight2

The second-smallest line height on a series of 8 steps.

LineHeight3

The base line height on a series of 8 steps.

LineHeight4

The fifth-largest line height on a series of 8 steps.

LineHeight5

The fourth-largest line height on a series of 8 steps.

LineHeight6

The third-largest line height on a series of 8 steps.

LineHeight7

The second-largest line height on a series of 8 steps.

LineHeight8

The largest line height on a series of 8 steps.

OpaqueWidgetColor

A [Color] to be used as a background color for widgets that render an opaque background.

OutlineColor

A [Color] to be used as an outline color.

SurfaceColor

The [Color] of the surface for the user interface to draw upon.

TextColor

The [Color] to use when rendering text.

TextColorVariant

The [Color] to use when rendering text in a more subdued tone.

TextSize

The Dimension to use as the size to render text.

TextSize1

The smallest text size on a series of 8 steps.

TextSize2

The second-smallest text size on a series of 8 steps.

TextSize3

The base text size on a series of 8 steps.

TextSize4

The fifth-largest text size on a series of 8 steps.

TextSize5

The fourth-largest text size on a series of 8 steps.

TextSize6

The third-largest text size on a series of 8 steps.

TextSize7

The second-largest text size on a series of 8 steps.

TextSize8

The largest text size on a series of 8 steps.

WidgetAccentColor

A [Color] to be used to accent a widget.

WidgetBackground

A [Color] to be used as the background color of a widget.