Trait gooey::styles::ComponentDefinition

pub trait ComponentDefinition: NamedComponent {
    type ComponentType: ComponentType;

    // Required method
    fn default_value(
        &self,
        context: &WidgetContext<'_, '_>
    ) -> Self::ComponentType;
}

A type that represents a named component with a default value of a specific Rust type.

Required Associated Types

type ComponentType: ComponentType

The type that will be contained in the Component.

Required Methods

fn default_value(&self, context: &WidgetContext<'_, '_>) -> Self::ComponentType

Returns the default value to use for this component.