Struct gooey::value::Dynamic

pub struct Dynamic<T>(/* private fields */);

An instance of a value that provides APIs to observe and react to its contents.

Implementations

impl<T> Dynamic<T>

pub fn new(value: T) -> Self

Creates a new instance wrapping value.

pub fn linked<R, TIntoR, TIntoRResult, RIntoT, RIntoTResult>( &self, t_into_r: TIntoR, r_into_t: RIntoT ) -> Dynamic<R>where T: PartialEq + Send + 'static, R: PartialEq + Send + 'static, TIntoRResult: Into<Option<R>> + Send + 'static, RIntoTResult: Into<Option<T>> + Send + 'static, TIntoR: FnMut(&T) -> TIntoRResult + Send + 'static, RIntoT: FnMut(&R) -> RIntoTResult + Send + 'static,

Returns a new dynamic that has its contents linked with self by the pair of mapping functions provided.

When the returned dynamic is updated, r_into_t will be invoked. This function accepts &R and can return T, or Option<T>. If a value is produced, self will be updated with the new value.

When self is updated, t_into_r will be invoked. This function accepts &T and can return R or Option<R>. If a value is produced, the returned dynamic will be updated with the new value.

Panics

This function panics if calling t_into_r with the current contents of the Dynamic produces a None value. This requirement is only for the first invocation, and it is guaranteed to occur before this function returns.

pub fn linked_string(&self) -> Dynamic<String>where T: ToString + FromStr + PartialEq + Send + 'static,

Creates a linked dynamic containing a String.

When self is updated, ToString::to_string() will be called to produce a new string value to store in the returned dynamic.

When the returned dynamic is updated, str::parse is called to produce a new T. If an error is returned, self will not be updated. Otherwise, self will be updated with the produced value.

pub fn map_ref<R>(&self, map: impl FnOnce(&T) -> R) -> R

Maps the contents with read-only access.

Panics

This function panics if this value is already locked by the current thread.

pub fn map_mut<R>(&self, map: impl FnOnce(&mut T) -> R) -> R

Maps the contents with exclusive access. Before returning from this function, all observers will be notified that the contents have been updated.

Panics

This function panics if this value is already locked by the current thread.

pub fn map_each_into<U>(&self) -> Dynamic<U>where U: From<T> + Send + 'static, T: Clone,

Returns a new dynamic that is updated using U::from(T.clone()) each time self is updated.

pub fn map_each_to<U>(&self) -> Dynamic<U>where U: for<'a> From<&'a T> + Send + 'static, T: Clone,

Returns a new dynamic that is updated using U::from(&T) each time self is updated.

pub fn for_each<F>(&self, for_each: F)where F: for<'a> FnMut(&'a T) + Send + 'static,

Attaches for_each to this value so that it is invoked each time the value’s contents are updated.

pub fn with_for_each<F>(self, for_each: F) -> Selfwhere F: for<'a> FnMut(&'a T) + Send + 'static,

Attaches for_each to this value so that it is invoked each time the value’s contents are updated. This function returns self.

pub fn map_each<R, F>(&self, map: F) -> Dynamic<R>where F: for<'a> FnMut(&'a T) -> R + Send + 'static, R: Send + 'static,

Creates a new dynamic value that contains the result of invoking map each time this value is changed.

pub fn map_each_unique<R, F>(&self, map: F) -> Dynamic<R>where F: for<'a> FnMut(&'a T) -> R + Send + 'static, R: Send + PartialEq + 'static,

Creates a new dynamic value that contains the result of invoking map each time this value is changed.

This version of map_each uses Dynamic::try_update to prevent deadlocks and debounce dependent values.

pub fn with_clone<R>(&self, with_clone: impl FnOnce(Self) -> R) -> R

A helper function that invokes with_clone with a clone of self. This code may produce slightly more readable code.

let value = gooey::value::Dynamic::new(1);

// Using with_clone
value.with_clone(|value| {
    std::thread::spawn(move || {
        println!("{}", value.get());
    })
});

// Using an explicit clone
std::thread::spawn({
    let value = value.clone();
    move || {
        println!("{}", value.get());
    }
});

println!("{}", value.get());

pub fn get(&self) -> Twhere T: Clone,

Returns a clone of the currently contained value.

Panics

This function panics if this value is already locked by the current thread.

pub fn get_tracking_refresh(&self, context: &WidgetContext<'_, '_>) -> Twhere T: Clone,

Returns a clone of the currently contained value.

context will be invalidated when the value is updated.

Panics

This function panics if this value is already locked by the current thread.

pub fn get_tracking_invalidate(&self, context: &WidgetContext<'_, '_>) -> Twhere T: Clone,

Returns a clone of the currently contained value.

context will be invalidated when the value is updated.

Panics

This function panics if this value is already locked by the current thread.

pub fn take(&self) -> Twhere T: Default,

Returns the currently stored value, replacing the current contents with T::default().

Panics

This function panics if this value is already locked by the current thread.

pub fn take_if_not_default(&self) -> Option<T>where T: Default + PartialEq,

Checks if the currently stored value is different than T::default(), and if so, returns Some(self.take()).

Panics

This function panics if this value is already locked by the current thread.

pub fn replace(&self, new_value: T) -> T

Replaces the contents with new_value, returning the previous contents. Before returning from this function, all observers will be notified that the contents have been updated.

Panics

This function panics if this value is already locked by the current thread.

pub fn set(&self, new_value: T)

Stores new_value in this dynamic. Before returning from this function, all observers will be notified that the contents have been updated.

Panics

This function panics if this value is already locked by the current thread.

pub fn update(&self, new_value: T) -> boolwhere T: PartialEq,

Updates this dynamic with new_value, but only if new_value is not equal to the currently stored value.

Returns true if the value was updated.

Panics

This function panics if this value is already locked by the current thread.

pub fn try_update(&self, new_value: T) -> Result<bool, T>where T: PartialEq,

Attempt to store new_value in self. If the value cannot be stored due to a deadlock, it is returned as an error.

Returns true if the value was updated.

pub fn create_reader(&self) -> DynamicReader<T>

Returns a new reference-based reader for this dynamic value.

Panics

This function panics if this value is already locked by the current thread.

pub fn into_reader(self) -> DynamicReader<T>

Converts this Dynamic into a reader.

Panics

This function panics if this value is already locked by the current thread.

pub fn lock(&self) -> DynamicGuard<'_, T>

Returns an exclusive reference to the contents of this dynamic.

This call will block until all other guards for this dynamic have been dropped.

Panics

This function panics if this value is already locked by the current thread.

pub fn generation(&self) -> Generation

Returns the current generation of the value.

Panics

This function panics if this value is already locked by the current thread.

pub fn transition_to(&self, new_value: T) -> DynamicTransition<T>where T: LinearInterpolate + Clone + Send + Sync,

Returns a pending transition for this value to new_value.

impl Dynamic<WidgetInstance>

pub fn switcher(self) -> Switcher

Returns a new Switcher widget whose contents is the value of this dynamic.