Struct gooey::widget::WidgetInstance

pub struct WidgetInstance { /* private fields */ }

An instance of a Widget.

Implementations

impl WidgetInstance

pub fn with_id<W>(widget: W, id: WidgetTag) -> Selfwhere W: Widget,

Returns a new instance containing widget that is assigned the unique id provided.

pub fn new<W>(widget: W) -> Selfwhere W: Widget,

Returns a new instance containing widget.

pub fn id(&self) -> WidgetId

Returns the unique id of this widget instance.

pub fn with_next_focus( self, next_focus: impl IntoValue<Option<WidgetId>> ) -> WidgetInstance

Sets the widget that should be focused next.

Gooey automatically determines reverse tab order by using this same relationship.

Panics

This function can only be called when one instance of the widget exists. If any clones exist, a panic will occur.

pub fn with_enabled(self, enabled: impl IntoValue<bool>) -> WidgetInstance

Sets this widget to be enabled/disabled based on enabled and returns self.

If this widget is disabled, all children widgets will also be disabled.

Panics

This function can only be called when one instance of the widget exists. If any clones exist, a panic will occur.

pub fn into_default(self) -> WidgetInstance

Sets this widget as a “default” widget.

Default widgets are automatically activated when the user signals they are ready for the default action to occur.

Example widgets this is used for are:

Submit buttons on forms

Ok buttons

Panics

This function can only be called when one instance of the widget exists. If any clones exist, a panic will occur.

pub fn into_escape(self) -> WidgetInstance

Sets this widget as an “escape” widget.

Escape widgets are automatically activated when the user signals they are ready to escape their current situation.

Example widgets this is used for are:

Close buttons

Cancel buttons

Panics

This function can only be called when one instance of the widget exists. If any clones exist, a panic will occur.

pub fn lock(&self) -> WidgetGuard<'_>

Locks the widget for exclusive access. Locking widgets should only be done for brief moments of time when you are certain no deadlocks can occur due to other widget locks being held.

pub fn run(self) -> Result

Runs this widget instance as an application.

pub fn next_focus(&self) -> Option<WidgetId>

Returns the id of the widget that should receive focus after this widget.

This value comes from MakeWidget::with_next_focus().

pub fn is_default(&self) -> bool

Returns true if this is a default widget.

See MakeWidget::into_default() for more information.

pub fn is_escape(&self) -> bool

Returns true if this is an escape widget.

See MakeWidget::into_escape() for more information.