Trait gooey::widget::MakeWidget

pub trait MakeWidget: Sized {

    // Required method
    fn make_widget(self) -> WidgetInstance;

    // Provided methods
    fn into_window(self) -> Window<WidgetInstance> { ... }
    fn with_styles(self, styles: impl Into<Styles>) -> Style
       where Self: Sized { ... }
    fn with(
        self,
        name: &impl NamedComponent,
        component: impl Into<Component>
    ) -> Style { ... }
    fn with_next_focus(
        self,
        next_focus: impl IntoValue<Option<WidgetId>>
    ) -> WidgetInstance { ... }
    fn into_default(self) -> WidgetInstance { ... }
    fn into_escape(self) -> WidgetInstance { ... }
    fn and(self, other: impl MakeWidget) -> Children { ... }
    fn expand(self) -> Expand { ... }
    fn expand_weighted(self, weight: u8) -> Expand { ... }
    fn centered(self) -> Align { ... }
    fn scroll(self) -> Scroll { ... }
    fn vertical_scroll(self) -> Scroll { ... }
    fn horizontal_scroll(self) -> Scroll { ... }

}

A type that can create a WidgetInstance.

Required Methods

fn make_widget(self) -> WidgetInstance

Returns a new widget.

Provided Methods

fn into_window(self) -> Window<WidgetInstance>

Returns a new window containing self as the root widget.

fn with_styles(self, styles: impl Into<Styles>) -> Stylewhere Self: Sized,

Associates styles with this widget.

This is equivalent to Style::new(styles, self).

fn with( self, name: &impl NamedComponent, component: impl Into<Component> ) -> Style

Associates a style component with self.

fn with_next_focus( self, next_focus: impl IntoValue<Option<WidgetId>> ) -> WidgetInstance

Sets the widget that should be focused next.

Gooey automatically determines reverse tab order by using this same relationship.

fn into_default(self) -> WidgetInstance

Sets this widget as a “default” widget.

Default widgets are automatically activated when the user signals they are ready for the default action to occur.

Example widgets this is used for are:

Submit buttons on forms

Ok buttons

fn into_escape(self) -> WidgetInstance

Sets this widget as an “escape” widget.

Escape widgets are automatically activated when the user signals they are ready to escape their current situation.

Example widgets this is used for are:

Close buttons

Cancel buttons

fn and(self, other: impl MakeWidget) -> Children

Returns a collection of widgets using self and other.

fn expand(self) -> Expand

Expands self to grow to fill its parent.

fn expand_weighted(self, weight: u8) -> Expand

Expands self to grow to fill its parent proportionally with other weighted siblings.

fn centered(self) -> Align

Aligns self to the center vertically and horizontally.

fn scroll(self) -> Scroll

Allows scrolling self both vertically and horizontally.

fn vertical_scroll(self) -> Scroll

Allows scrolling self vertically.

fn horizontal_scroll(self) -> Scroll

Allows scrolling self horizontally.