Trait gooey::widget::Widget

pub trait Widget: Send + UnwindSafe + Debug + 'static {

    // Required methods
    fn redraw(&mut self, context: &mut GraphicsContext<'_, '_, '_, '_, '_>);
    fn layout(
        &mut self,
        available_space: Size<ConstraintLimit>,
        context: &mut LayoutContext<'_, '_, '_, '_, '_>
    ) -> Size<UPx>;

    // Provided methods
    fn mounted(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn unmounted(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn hit_test(
        &mut self,
        location: Point<Px>,
        context: &mut EventContext<'_, '_>
    ) -> bool { ... }
    fn hover(&mut self, location: Point<Px>, context: &mut EventContext<'_, '_>) { ... }
    fn unhover(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn accept_focus(&mut self, context: &mut EventContext<'_, '_>) -> bool { ... }
    fn focus(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn blur(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn activate(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn deactivate(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn mouse_down(
        &mut self,
        location: Point<Px>,
        device_id: DeviceId,
        button: MouseButton,
        context: &mut EventContext<'_, '_>
    ) -> EventHandling { ... }
    fn mouse_drag(
        &mut self,
        location: Point<Px>,
        device_id: DeviceId,
        button: MouseButton,
        context: &mut EventContext<'_, '_>
    ) { ... }
    fn mouse_up(
        &mut self,
        location: Option<Point<Px>>,
        device_id: DeviceId,
        button: MouseButton,
        context: &mut EventContext<'_, '_>
    ) { ... }
    fn keyboard_input(
        &mut self,
        device_id: DeviceId,
        input: KeyEvent,
        is_synthetic: bool,
        context: &mut EventContext<'_, '_>
    ) -> EventHandling { ... }
    fn ime(
        &mut self,
        ime: Ime,
        context: &mut EventContext<'_, '_>
    ) -> EventHandling { ... }
    fn mouse_wheel(
        &mut self,
        device_id: DeviceId,
        delta: MouseScrollDelta,
        phase: TouchPhase,
        context: &mut EventContext<'_, '_>
    ) -> EventHandling { ... }

}

A type that makes up a graphical user interface.

This type can go by many names in other UI frameworks: View, Component, Control.

Required Methods

fn redraw(&mut self, context: &mut GraphicsContext<'_, '_, '_, '_, '_>)

Redraw the contents of this widget.

fn layout( &mut self, available_space: Size<ConstraintLimit>, context: &mut LayoutContext<'_, '_, '_, '_, '_> ) -> Size<UPx>

Layout this widget and returns the ideal size based on its contents and the available_space.

Provided Methods

fn mounted(&mut self, context: &mut EventContext<'_, '_>)

The widget has been mounted into a parent widget.

fn unmounted(&mut self, context: &mut EventContext<'_, '_>)

The widget has been removed from its parent widget.

fn hit_test( &mut self, location: Point<Px>, context: &mut EventContext<'_, '_> ) -> bool

Returns true if this widget should respond to mouse input at location.

fn hover(&mut self, location: Point<Px>, context: &mut EventContext<'_, '_>)

The widget is currently has a cursor hovering it at location.

fn unhover(&mut self, context: &mut EventContext<'_, '_>)

The widget is no longer being hovered.

fn accept_focus(&mut self, context: &mut EventContext<'_, '_>) -> bool

This widget has been targeted to be focused. If this function returns true, the widget will be focused. If false, Gooey will continue searching for another focus target.

fn focus(&mut self, context: &mut EventContext<'_, '_>)

The widget has received focus for user input.

fn blur(&mut self, context: &mut EventContext<'_, '_>)

The widget is no longer focused for user input.

fn activate(&mut self, context: &mut EventContext<'_, '_>)

The widget has become the active widget.

fn deactivate(&mut self, context: &mut EventContext<'_, '_>)

The widget is no longer active.

fn mouse_down( &mut self, location: Point<Px>, device_id: DeviceId, button: MouseButton, context: &mut EventContext<'_, '_> ) -> EventHandling

A mouse button event has occurred at location. Returns whether the event has been handled or not.

If an event is handled, the widget will receive callbacks for mouse_drag and mouse_up.

fn mouse_drag( &mut self, location: Point<Px>, device_id: DeviceId, button: MouseButton, context: &mut EventContext<'_, '_> )

A mouse button is being held down as the cursor is moved across the widget.

fn mouse_up( &mut self, location: Option<Point<Px>>, device_id: DeviceId, button: MouseButton, context: &mut EventContext<'_, '_> )

A mouse button is no longer being pressed.

fn keyboard_input( &mut self, device_id: DeviceId, input: KeyEvent, is_synthetic: bool, context: &mut EventContext<'_, '_> ) -> EventHandling

A keyboard event has been sent to this widget. Returns whether the event has been handled or not.

fn ime(&mut self, ime: Ime, context: &mut EventContext<'_, '_>) -> EventHandling

An input manager event has been sent to this widget. Returns whether the event has been handled or not.

fn mouse_wheel( &mut self, device_id: DeviceId, delta: MouseScrollDelta, phase: TouchPhase, context: &mut EventContext<'_, '_> ) -> EventHandling

A mouse wheel event has been sent to this widget. Returns whether the event has been handled or not.