Trait gooey::widget::Widget

pub trait Widget: Send + UnwindSafe + Debug + 'static {

    // Required method
    fn redraw(&mut self, context: &mut GraphicsContext<'_, '_, '_, '_, '_>);

    // Provided methods
    fn summarize(&self, f: &mut Formatter<'_>) -> Result { ... }
    fn full_control_redraw(&self) -> bool { ... }
    fn layout(
        &mut self,
        available_space: Size<ConstraintLimit>,
        context: &mut LayoutContext<'_, '_, '_, '_, '_>
    ) -> Size<UPx> { ... }
    fn mounted(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn unmounted(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn hit_test(
        &mut self,
        location: Point<Px>,
        context: &mut EventContext<'_, '_>
    ) -> bool { ... }
    fn hover(
        &mut self,
        location: Point<Px>,
        context: &mut EventContext<'_, '_>
    ) -> Option<CursorIcon> { ... }
    fn unhover(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn accept_focus(&mut self, context: &mut EventContext<'_, '_>) -> bool { ... }
    fn focus(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn advance_focus(
        &mut self,
        direction: VisualOrder,
        context: &mut EventContext<'_, '_>
    ) -> EventHandling { ... }
    fn allow_blur(&mut self, context: &mut EventContext<'_, '_>) -> bool { ... }
    fn blur(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn activate(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn deactivate(&mut self, context: &mut EventContext<'_, '_>) { ... }
    fn mouse_down(
        &mut self,
        location: Point<Px>,
        device_id: DeviceId,
        button: MouseButton,
        context: &mut EventContext<'_, '_>
    ) -> EventHandling { ... }
    fn mouse_drag(
        &mut self,
        location: Point<Px>,
        device_id: DeviceId,
        button: MouseButton,
        context: &mut EventContext<'_, '_>
    ) { ... }
    fn mouse_up(
        &mut self,
        location: Option<Point<Px>>,
        device_id: DeviceId,
        button: MouseButton,
        context: &mut EventContext<'_, '_>
    ) { ... }
    fn keyboard_input(
        &mut self,
        device_id: DeviceId,
        input: KeyEvent,
        is_synthetic: bool,
        context: &mut EventContext<'_, '_>
    ) -> EventHandling { ... }
    fn ime(
        &mut self,
        ime: Ime,
        context: &mut EventContext<'_, '_>
    ) -> EventHandling { ... }
    fn mouse_wheel(
        &mut self,
        device_id: DeviceId,
        delta: MouseScrollDelta,
        phase: TouchPhase,
        context: &mut EventContext<'_, '_>
    ) -> EventHandling { ... }
    fn root_behavior(
        &mut self,
        context: &mut EventContext<'_, '_>
    ) -> Option<(RootBehavior, WidgetInstance)> { ... }

}

A type that makes up a graphical user interface.

This type can go by many names in other UI frameworks: View, Component, Control.

Required Methods

fn redraw(&mut self, context: &mut GraphicsContext<'_, '_, '_, '_, '_>)

Redraw the contents of this widget.

Provided Methods

fn summarize(&self, f: &mut Formatter<'_>) -> Result

Writes a summary of this widget into fmt.

The default implementation calls Debug::fmt. This function allows widget authors to print only publicly relevant information that will appear when debug formatting a WidgetInstance.

fn full_control_redraw(&self) -> bool

Returns true if this widget handles all built-in style components that apply.

These components are:

Opacity

WidgetBackground

FontFamily

TextSize

LineHeight

FontStyle

FontWeight

fn layout( &mut self, available_space: Size<ConstraintLimit>, context: &mut LayoutContext<'_, '_, '_, '_, '_> ) -> Size<UPx>

Layout this widget and returns the ideal size based on its contents and the available_space.

fn mounted(&mut self, context: &mut EventContext<'_, '_>)

The widget has been mounted into a parent widget.

fn unmounted(&mut self, context: &mut EventContext<'_, '_>)

The widget has been removed from its parent widget.

fn hit_test( &mut self, location: Point<Px>, context: &mut EventContext<'_, '_> ) -> bool

Returns true if this widget should respond to mouse input at location.

fn hover( &mut self, location: Point<Px>, context: &mut EventContext<'_, '_> ) -> Option<CursorIcon>

The widget is currently has a cursor hovering it at location.

fn unhover(&mut self, context: &mut EventContext<'_, '_>)

The widget is no longer being hovered.

fn accept_focus(&mut self, context: &mut EventContext<'_, '_>) -> bool

This widget has been targeted to be focused. If this function returns true, the widget will be focused. If false, Gooey will continue searching for another focus target.

fn focus(&mut self, context: &mut EventContext<'_, '_>)

The widget has received focus for user input.

fn advance_focus( &mut self, direction: VisualOrder, context: &mut EventContext<'_, '_> ) -> EventHandling

The widget should switch to the next focusable area within this widget, honoring direction in a consistent manner. Returning HANDLED will cause the search for the next focus widget stop.

fn allow_blur(&mut self, context: &mut EventContext<'_, '_>) -> bool

The widget is about to lose focus. Returning true allows the focus to switch away from this widget.

fn blur(&mut self, context: &mut EventContext<'_, '_>)

The widget is no longer focused for user input.

fn activate(&mut self, context: &mut EventContext<'_, '_>)

The widget has become the active widget.

fn deactivate(&mut self, context: &mut EventContext<'_, '_>)

The widget is no longer active.

fn mouse_down( &mut self, location: Point<Px>, device_id: DeviceId, button: MouseButton, context: &mut EventContext<'_, '_> ) -> EventHandling

A mouse button event has occurred at location. Returns whether the event has been handled or not.

If an event is handled, the widget will receive callbacks for mouse_drag and mouse_up.

fn mouse_drag( &mut self, location: Point<Px>, device_id: DeviceId, button: MouseButton, context: &mut EventContext<'_, '_> )

A mouse button is being held down as the cursor is moved across the widget.

fn mouse_up( &mut self, location: Option<Point<Px>>, device_id: DeviceId, button: MouseButton, context: &mut EventContext<'_, '_> )

A mouse button is no longer being pressed.

fn keyboard_input( &mut self, device_id: DeviceId, input: KeyEvent, is_synthetic: bool, context: &mut EventContext<'_, '_> ) -> EventHandling

A keyboard event has been sent to this widget. Returns whether the event has been handled or not.

fn ime(&mut self, ime: Ime, context: &mut EventContext<'_, '_>) -> EventHandling

An input manager event has been sent to this widget. Returns whether the event has been handled or not.

fn mouse_wheel( &mut self, device_id: DeviceId, delta: MouseScrollDelta, phase: TouchPhase, context: &mut EventContext<'_, '_> ) -> EventHandling

A mouse wheel event has been sent to this widget. Returns whether the event has been handled or not.

fn root_behavior( &mut self, context: &mut EventContext<'_, '_> ) -> Option<(RootBehavior, WidgetInstance)>

Returns a reference to a single child widget if this widget is a widget that primarily wraps a single other widget to customize its behavior.